### 2026-07-24 92c4f49
Fix formatting in DESCRIPTION file
Removed a line break in the DESCRIPTION file to improve formatting.
