+++ DESCRIPTION
@@ -5,8 +5,8 @@ Authors@R:
     person("Joash Joshua", "Ayo", , "joashjoshua789@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0009-0007-1642-0172"))
 Description: Provides pipe-friendly, stateless access to every endpoint in the 
-    Breeding API (BrAPI) v2.1 specification. Covers all four BrAPI modules — 
-    Core, Germplasm, Phenotyping, and Genotyping — returning tidy tibbles ready 
+    Breeding API (BrAPI) v2.1 specification. Covers all four BrAPI modules 
+    Core, Germplasm, Phenotyping, and Genotyping returning tidy tibbles ready 
     for analysis. Features include automatic pagination, async search handling, 
     response caching, parallel batch fetching, and convenience functions for 
     genomic selection workflows (e.g. dosage matrix extraction). Designed for 
