+++ reference/brapi_search_calls.md
@@ -38,7 +38,7 @@ A tibble of matching genotype calls.
 # \donttest{
 con <- brapi_connection("https://test-server.brapi.org")
 brapi_search_calls(con, variantSetDbIds = "variantset1")
-#> ℹ Async search started (ID: 4d9d23a7-56b6-4ffa-8499-8de30a43bb5c). Polling...
+#> ℹ Async search started (ID: 6b87e0e8-5ce0-4313-b89d-a42988f6f379). Polling...
 #> # A tibble: 260 × 12
 #>    additionalInfo callSetDbId callSetName genotype         genotypeValue
 #>    <lgl>          <chr>       <chr>       <list>           <chr>        
