+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   genomic-selection-pipeline: genomic-selection-pipeline.html
   getting-started: getting-started.html
   multi-environment-analysis: multi-environment-analysis.html
-last_built: 2026-07-15T12:37Z
+last_built: 2026-07-24T19:37Z
 urls:
   reference: https://josh45-source.github.io/brapiR2/reference
   article: https://josh45-source.github.io/brapiR2/articles
