### 2026-08-08 35d4abd
docs(EC5, EC7): credential guidance, identify BrAPI in DESCRIPTION
Adds keyring to Suggests and cross-references the new credential
handling section from brapi_login() and brapi_set_token().

DESCRIPTION now names the issuing organisation and gives the public URL
per packaging guide section 1.7.1, single-quotes external software
names, and consolidates article-only dependencies in
Config/Needs/website.
