### 2026-08-08 148e792
fix: resolve studyDbId against .env in brapi_study_data() fallback
The client-side fallback filtered with .data\ == studyDbId.
Because the function argument shares the column name, dplyr's data mask
resolved the bare symbol to the column, making the comparison always
TRUE and silently returning the first study's observations for any
requested study. Adds a regression test that fails against the old
behaviour.
