+++ tests/testthat/test-phenotyping-mocked.R
@@ -61,6 +61,36 @@ test_that("brapi_study_data falls back to client-side filtering", {
   expect_identical(wide$Yield, "10")
 })
 
+test_that("brapi_study_data fallback filters by requested study, not itself", {
+  # Regression test: the fallback previously did
+  # `dplyr::filter(all_obs, .data$studyDbId == studyDbId)`, where the bare
+  # `studyDbId` inside filter() resolves to the DATA COLUMN (same name),
+  # not the function argument - making the comparison a tautology that
+  # matched every row regardless of which study was requested.
+  local_mocked_bindings(
+    brapi_observations = function(con, studyDbId = NULL, ...) {
+      if (!is.null(studyDbId)) {
+        tibble::tibble()
+      } else {
+        tibble::tibble(
+          observationUnitDbId    = "ou1",
+          studyDbId               = "study1",
+          observationVariableName = "Yield",
+          value                   = "10"
+        )
+      }
+    },
+    .package = "brapiR2"
+  )
+
+  con <- brapi_connection("https://example.org")
+
+  # A study genuinely absent from the unfiltered data must come back empty,
+  # not silently return study1's rows.
+  wide <- suppressMessages(brapi_study_data(con, "study2"))
+  expect_identical(nrow(wide), 0L)
+})
+
 test_that("brapi_study_data uses observationValue/DbId fallback column names", {
   local_mocked_bindings(
     brapi_observations = function(con, studyDbId = NULL, ...) {
@@ -79,4 +109,6 @@ test_that("brapi_study_data uses observationValue/DbId fallback column names", {
 
   expect_identical(nrow(wide), 1L)
   expect_true(all(c("var1", "var2") %in% names(wide)))
+  expect_identical(wide$var1, "10")
+  expect_identical(wide$var2, "150")
 })
