### 2026-08-08 08cc1db
test: assert on captured arguments rather than shape
The thin-wrapper mocks returned a constant regardless of input, so
roughly 40 assertions could not have detected a wrapper calling the
wrong endpoint or dropping an ID. Mocks now capture endpoint, query and
body, and each wrapper is checked against its BrAPI path. Verified by
deliberately breaking an endpoint and a filter argument; both caught.
