+++ tests/testthat/test-thin-wrappers-mocked.R
@@ -1,82 +1,235 @@
 ## Most exported list/get functions are one-line wrappers around
 ## brapi_get() or brapi_post_search(). Mocking those two internal
 ## functions once exercises every wrapper's own body without any network.
+##
+## The mocks below capture the arguments they were called with (endpoint,
+## query, body) so each assertion can check that a wrapper hit the right
+## BrAPI path and threaded its ID/filter arguments through under the
+## correct key - not just that it returned whatever the mock handed back.
 
-test_that("GET-based list/detail wrappers call brapi_get() and return it", {
+test_that("core wrappers hit the right endpoint and pass filters through", {
+  captured <- new.env()
   canned <- tibble::tibble(id = "x")
   local_mocked_bindings(
-    brapi_get = function(con, endpoint, query = list()) canned,
+    brapi_get = function(con, endpoint, query = list()) {
+      captured$endpoint <- endpoint
+      captured$query <- query
+      canned
+    },
     .package = "brapiR2"
   )
 
   con <- brapi_connection("https://example.org")
 
   expect_identical(brapi_programs(con), canned)
+  expect_identical(captured$endpoint, "/programs")
+  brapi_programs(con, commonCropName = "rice")
+  expect_identical(captured$query$commonCropName, "rice")
+
   expect_identical(brapi_program(con, "p1"), canned)
+  expect_identical(captured$endpoint, "/programs/p1")
+
   expect_identical(brapi_trials(con), canned)
-  expect_identical(brapi_trials(con, programDbId = "p1"), canned)
+  expect_identical(captured$endpoint, "/trials")
+  brapi_trials(con, programDbId = "p1")
+  expect_identical(captured$query$programDbId, "p1")
+
   expect_identical(brapi_trial(con, "t1"), canned)
+  expect_identical(captured$endpoint, "/trials/t1")
+
   expect_identical(brapi_studies(con), canned)
-  expect_identical(brapi_studies(con, trialDbId = "t1"), canned)
+  expect_identical(captured$endpoint, "/studies")
+  brapi_studies(con, trialDbId = "t1")
+  expect_identical(captured$query$trialDbId, "t1")
+
   expect_identical(brapi_study(con, "s1"), canned)
+  expect_identical(captured$endpoint, "/studies/s1")
+
   expect_identical(brapi_locations(con), canned)
+  expect_identical(captured$endpoint, "/locations")
+  brapi_locations(con, locationType = "Field")
+  expect_identical(captured$query$locationType, "Field")
+
   expect_identical(brapi_seasons(con), canned)
+  expect_identical(captured$endpoint, "/seasons")
+  brapi_seasons(con, year = 2020)
+  expect_identical(captured$query$year, 2020)
+
   expect_identical(brapi_lists(con), canned)
+  expect_identical(captured$endpoint, "/lists")
+
   expect_identical(brapi_people(con), canned)
-
-  expect_identical(brapi_samples(con), canned)
-  expect_identical(brapi_variants(con), canned)
-  expect_identical(brapi_variants(con, variantSetDbId = "vs1"), canned)
-  expect_identical(brapi_variant_sets(con), canned)
-  expect_identical(brapi_variant_sets(con, studyDbId = "s1"), canned)
-  expect_identical(brapi_calls(con), canned)
-  expect_identical(brapi_calls(con, variantSetDbId = "vs1"), canned)
-  expect_identical(brapi_call_sets(con), canned)
-  expect_identical(brapi_references(con), canned)
-  expect_identical(brapi_reference_sets(con), canned)
-
-  expect_identical(brapi_germplasm(con), canned)
[…]
