+++ tests/testthat/test-cache-integration.R
@@ -2,7 +2,7 @@
 ## Require network access to test-server.brapi.org; skipped on CRAN and offline.
 
 skip_if_offline_brapi <- function() {
-  skip_if_offline(host = "test-server.brapi.org")
+  testthat::skip_if_offline(host = "test-server.brapi.org")
 }
 
 SERVER <- "https://test-server.brapi.org"
