### 2026-08-08 0c6a353
fix(EC10): respect the caller's future plan in brapi_fetch_parallel()
The function set plan(multisession) internally and restored the previous
plan on exit. The cleanup was correct, but setting a plan inside a
package function mutates session-wide state the caller did not ask for,
silently replaces a backend they may have configured deliberately, and
forces a parallel toggle to propagate to anyone calling into this
package.

It now uses whatever plan is active, defaulting to sequential. .workers
is deprecated rather than removed, warning and pointing to
future::plan(). Documentation, both call
[…]
