### 2026-08-08 0e22532
docs(EC1, EC2, EC3): evaluate live output, add explanatory prose
Renames the introductory vignette to brapiR2.Rmd so pkgdown promotes it
to Get started. Moves the two workflow documents to vignettes/articles/.

Every brapiR2 call that runs against the public test server now
evaluates, guarded by a reachability check so the build survives an
unreachable server. Only credentialed calls remain illustrative. The
genomic selection article shows the phenotype/genotype overlap live as
evidence for why the modelling step uses simulated data.

Replaces the independent-marker simulation with AlphaSimR, 
[…]
