+++ dev/explore-test-server.R
@@ -1,0 +1,337 @@
+# Exploratory script: does the public BrAPI test server hold enough data
+# to support an evaluated genomic selection article?
+#
+# Not run as part of R CMD check or the test suite. Run interactively with:
+#   Rscript dev/explore-test-server.R
+#
+# Answers, in order:
+#   1. Phenotype side  - studies, observation units/observations, variables
+#   2. Genotype side   - variant sets, variants, call sets, dosage matrix
+#   3. Overlap         - how many germplasm have BOTH phenotype and genotype data
+#   4. Verdict         - paired germplasm count, marker count, trait variance
+
+library(brapiR2)
+suppressPackageStartupMessages({
+  library(dplyr)
+})
+
+SERVER <- "https://test-server.brapi.org"
+
+section <- function(title) {
+  cat("\n", strrep("=", 70), "\n", sep = "")
+  cat(title, "\n")
+  cat(strrep("=", 70), "\n", sep = "")
+}
+
+safely <- function(expr, label) {
+  tryCatch(
+    expr,
+    error = function(e) {
+      cat(sprintf("  [ERROR] %s: %s\n", label, conditionMessage(e)))
+      NULL
+    }
+  )
+}
+
+con <- brapi_connection(SERVER)
+
+section("CONNECTIVITY")
+reachable <- safely(brapi_ping(con), "brapi_ping")
+if (!isTRUE(reachable)) {
+  cat("Server is not reachable. Aborting exploration.\n")
+  quit(save = "no", status = 1L)
+}
+
+## ------------------------------------------------------------------------
+## 1. PHENOTYPE SIDE
+## ------------------------------------------------------------------------
+section("1. PHENOTYPE SIDE")
+
+studies <- safely(brapi_studies(con), "brapi_studies")
+n_studies <- if (is.null(studies)) 0L else nrow(studies)
+cat(sprintf("Studies available: %d\n", n_studies))
+
+study_summary <- NULL
+if (n_studies > 0L) {
+  n_preview <- min(5L, n_studies)
+  cat(sprintf("\nInspecting the first %d studies:\n", n_preview))
+
+  study_summary <- lapply(seq_len(n_preview), function(i) {
+    sid <- studies$studyDbId[i]
+    sname <- if ("studyName" %in% names(studies)) studies$studyName[i] else NA_character_
+
+    ou <- safely(brapi_observation_units(con, studyDbId = sid), "brapi_observation_units")
+    obs <- safely(brapi_observations(con, studyDbId = sid), "brapi_observations")
+
+    n_ou <- if (is.null(ou)) NA_integer_ else nrow(ou)
+    n_obs <- if (is.null(obs)) NA_integer_ else nrow(obs)
+
+    cat(sprintf(
+      "  - %s (%s): %s observation units, %s observations\n",
+      sid, if (is.null(sname) || is.na(sname)) "?" else sname,
+      ifelse(is.na(n_ou), "ERROR", n_ou),
+      ifelse(is.na(n_obs), "ERROR", n_obs)
+    ))
+
+    tibble(studyDbId = sid, studyName = sname, n_ou = n_ou, n_obs = n_obs)
+  })
+  study_summary <- bind_rows(study_summary)
+}
+
+vars <- safely(brapi_observation_variables(con), "brapi_observation_variables")
+n_vars <- if (is.null(vars)) 0L else nrow(vars)
+cat(sprintf("\nObservation variables defined on server: %d\n", n_vars))
+if (n_vars > 0L) {
+  name_col <- intersect(c("observationVariableName", "name"), names(vars))[1]
+  if (!is.na(name_col) && !is.null(name_col)) {
+    cat("Examples:", paste(utils::head(vars[[name_col]], 10), collapse = ", "), "\n")
+  }
+}
+
[…]
