+++ vignettes/articles/genomic-selection-pipeline.Rmd
@@ -1,0 +1,567 @@
+---
+title: "Genomic Selection Pipeline with brapiR2"
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
+con <- brapiR2::brapi_connection("https://test-server.brapi.org")
+server_up <- isTRUE(tryCatch(
+  brapiR2::brapi_ping(con),
+  error = function(e) FALSE
+))
+```
+
+```{r server-down-notice, eval = !server_up, echo = FALSE, results = "asis"}
+cat(
+  "> **Note:** the public BrAPI test server",
+  "(`https://test-server.brapi.org`) was unreachable when this article was",
+  "built, so Part 1's live output below was skipped. Part 3 (simulated",
+  "data) is unaffected and still ran."
+)
+```
+
+## Overview
+
+This article demonstrates a complete genomic selection (GS) workflow using
+brapiR2 to pull phenotypic and genotypic data from a BrAPI-compliant server
+into R, then running GS models with rrBLUP, BGLR, sommer, and AGHmatrix.
+
+The `brapi_get_dosage_matrix()` function returns a standard numeric matrix
+(samples x markers) that is directly compatible with the major GS packages:
+
+| Package | Usage |
+|---------|-------|
+| [rrBLUP](https://cran.r-project.org/package=rrBLUP) | `mixed.solve(y, Z = dosage)` |
+| [BGLR](https://cran.r-project.org/package=BGLR) | `ETA = list(MRK = list(X = dosage, model = "BRR"))` |
+| [sommer](https://cran.r-project.org/package=sommer) | `mmer(yield ~ 1, random = ~ vsr(germplasmName, Gu = A))` |
+| [AGHmatrix](https://cran.r-project.org/package=AGHmatrix) | `Gmatrix(dosage, method = "VanRaden")` |
+
+It has three parts: **Part 1** fetches real data from the public BrAPI test
+server and shows exactly what brapiR2 returns. **Part 2** shows the
+authenticated pattern you'd use against your own server, without running it.
+**Part 3** runs the actual GS modeling on a simulated dataset, because (as
+Part 1 demonstrates directly) the public server doesn't have enough real
+individuals with both phenotype and genotype data to fit and validate a
+model.
+
+## Part 1: Live Data from the Public Test Server
+
+These calls run against `https://test-server.brapi.org` and require no
+authentication.
+
+```{r pheno-live, eval = server_up}
+library(brapiR2)
+library(dplyr)
+
+studies <- brapi_studies(con)
+studies
+
+# brapi_study_data() falls back to fetching all observations and filtering
+# client-side when a server (like this one) doesn't honour the studyDbId
+# filter on /observations - so it succeeds even though a direct
+# brapi_observations(con, studyDbId = ...) call would come back empty.
+pheno_by_study <- lapply(studies$studyDbId, brapi_study_data, con = con)
+names(pheno_by_study) <- studies$studyDbId
+pheno_by_study[["study1"]]
+```
+
+```{r geno-live, eval = server_up}
+vsets <- brapi_variant_sets(con)
+vsets
+
+vs_id <- vsets$variantSetDbId[1]
+
+markers <- brapi_get_marker_map(con, vs_id)
+markers # referenceName/start are NA on this server; identity columns are real
+
+dosage <- brapi_get_dosage_matrix(con, vs_id)
+dim(dosage)
+dosage[seq_len(min(5, nrow(dosage))), seq_len(min(6, ncol(dosage)))]
+```
+
+### Do phenotype and genotype records share any germplasm here?
+
+This is the question that decides whether real genomic prediction is
[…]

[…]
