### 2026-08-08 7b82e05
style(EC9, EC11): single-line importFrom, clear lints and goodpractice
Splits multi-line @importFrom tags, which roxygen2 rejects.

Resolves lintr findings, replacing superassignment in test helpers with
environment-based state rather than removing the shared state they
need. Adds .lintr disabling object_name_linter package-wide, since
BrAPI field names are camelCase by specification.

Consolidates duplicated @param blocks into shared roxygen topics, and
adds documentation spelling fixes with domain vocabulary in
inst/WORDLIST.
