### 2026-08-08 3d6770b
docs(EC6): correct NAMESPACE guidance, document test setup
Contributors were told to edit NAMESPACE by hand, contradicting its
roxygen2 header. Now directs them to @export plus devtools::document().
Adds a section covering what is needed to run the test suite.
