+++ articles/multi-environment-analysis.md
@@ -2,142 +2,267 @@
 
 ## Overview
 
-Multi-environment trial (MET) analysis is the core workflow of a plant
+Multi-environment trial (MET) analysis is a core workflow of a plant
 breeding program: gather phenotypic observations from many locations and
-years, then decompose genotype × environment (GxE) interaction to
+years, then decompose genotype x environment (GxE) interaction to
 identify stable, high-performing varieties.
 
-This vignette shows how to:
+This article has three parts: **Part 1** fetches real study, location,
+and phenotype data from the public BrAPI test server and shows exactly
+what brapiR2 returns. **Part 2** shows the authenticated, parallel-fetch
+pattern you’d use against your own server, without running it. **Part
+3** runs the actual GxE analysis — AMMI/GGE biplots (**metan**) and
+mixed-model stability (**lme4**) — on a simulated trial, because Part
+1’s real data is nowhere near large enough for either method.
 
-1.  Pull phenotypic data from multiple BrAPI studies in parallel
-2.  Attach environment metadata (location, season)
-3.  Run AMMI / GGE biplots with the **metan** package
-4.  Run mixed-model stability analysis with **lme4**
+## Part 1: Live Data from the Public Test Server
 
-## Step 1: Connect and Discover Studies
+The three calls below return the pieces a multi-environment analysis is
+assembled from. **Studies** are the environments themselves: one trial
+grown at one site in one season. **Locations** carry the site metadata,
+latitude, longitude, altitude and so on, that lets you group
+environments sensibly or relate performance to conditions. **Seasons**
+identify the planting cycle, which for maize may be a single year in
+temperate programs or distinct long and short rains in East Africa.
+
+The distinction between a location and an environment is worth being
+precise about, because they are routinely conflated. An environment is a
+location and season combination. The same site in two seasons
+contributes two environments, and often two quite different ones, since
+rainfall and temperature in a given year can matter more for yield than
+the site itself. This is why a study, not a location, is the unit an
+AMMI or GGE analysis treats as an environment.
 
 ``` r
 
 library(brapiR2)
 library(dplyr)
+#> 
+#> Attaching package: 'dplyr'
+#> The following objects are masked from 'package:stats':
+#> 
+#>     filter, lag
+#> The following objects are masked from 'package:base':
+#> 
+#>     intersect, setdiff, setequal, union
 
-con <- brapi_connection("https://my-breedbase.org")
-con <- brapi_login(con, "username", "password")
+studies <- brapi_studies(con)
+studies
+#> # A tibble: 3 × 29
+#>   additionalInfo   externalReferences active commonCropName contacts  
+#>   <list>           <list>             <lgl>  <chr>          <list>    
+#> 1 <named list [1]> <list [1]>         TRUE   Tomatillo      <list [1]>
+#> 2 <named list [1]> <list [1]>         TRUE   Tomatillo      <list [1]>
+#> 3 <named list [1]> <list [1]>         TRUE   Tomatillo      <list [1]>
+#> # ℹ 24 more variables: culturalPractices <chr>, dataLinks <list>,
+#> #   documentationURL <chr>, endDate <chr>, environmentParameters <list>,
+#> #   experimentalDesign <list>, growthFacility <list>, lastUpdate <list>,
[…]
