+++ reference/brapi_fetch_parallel.md
@@ -7,7 +7,7 @@ or germplasm records simultaneously.
 ## Usage
 
 ``` r
-brapi_fetch_parallel(con, .fn, ids, .workers = 4L, ...)
+brapi_fetch_parallel(con, .fn, ids, .workers = NULL, ...)
 ```
 
 ## Arguments
@@ -28,7 +28,11 @@ brapi_fetch_parallel(con, .fn, ids, .workers = 4L, ...)
 
 - .workers:
 
-  Integer. Number of parallel workers. Default 4.
+  Deprecated. No longer used - the parallel backend is now the caller's
+  choice, set via
+  [`future::plan()`](https://future.futureverse.org/reference/plan.html)
+  before calling this function. Supplying a non-`NULL` value emits a
+  deprecation warning and otherwise has no effect.
 
 - ...:
 
@@ -38,14 +42,38 @@ brapi_fetch_parallel(con, .fn, ids, .workers = 4L, ...)
 
 A tibble with results from all IDs combined.
 
+## Details
+
+This function uses whatever `future` plan is already active when it is
+called, and does not set or restore one itself. If you have not called
+[`future::plan()`](https://future.futureverse.org/reference/plan.html),
+[`furrr::future_map_dfr()`](https://furrr.futureverse.org/reference/future_map.html)
+falls back to
+[`future::sequential`](https://future.futureverse.org/reference/sequential.html),
+so nothing runs in parallel until you set a plan yourself - call
+`future::plan(future::multisession, workers = N)` before this function
+to fetch in parallel, and `future::plan(future::sequential)` afterwards
+to shut the workers back down. Per the future package's best-practices
+vignette, choosing the parallel backend is the caller's decision: a
+package that sets and restores a plan on every call still mutates
+session-wide state the caller did not ask it to touch, and can silently
+replace a backend they configured deliberately.
+
 ## Examples
 
 ``` r
 # \donttest{
 con <- brapi_connection("https://test-server.brapi.org")
 study_ids <- c("study1", "study2", "study3")
+
+# Set the parallel backend yourself before calling; brapi_fetch_parallel()
+# uses whatever plan is active rather than setting one for you.
+future::plan(future::multisession, workers = 2)
 all_data <- brapi_fetch_parallel(con, brapi_study_data, study_ids)
-#> ℹ Fetching 3 items across 4 workers...
-#> ✔ Fetched 6 rows from 3 sources.
+#> ℹ Fetching 3 items...
+#> ! No observations found for study "study2".
+#> ! No observations found for study "study3".
+#> ✔ Fetched 2 rows from 3 sources.
+future::plan(future::sequential) # shut the workers back down when done
 # }
 ```
