+++ articles/index.md
@@ -2,9 +2,8 @@
 
 ### All vignettes
 
+- [brapiR2](https://josh45-source.github.io/brapiR2/articles/brapiR2.md):
 - [Genomic Selection Pipeline with
   brapiR2](https://josh45-source.github.io/brapiR2/articles/genomic-selection-pipeline.md):
-- [Getting Started with
-  brapiR2](https://josh45-source.github.io/brapiR2/articles/getting-started.md):
 - [Multi-Environment Trial (MET) Analysis with
   brapiR2](https://josh45-source.github.io/brapiR2/articles/multi-environment-analysis.md):

[…]
