### 2026-08-08 7b0eb69
ci: mark workspace safe for git in pkgcheck workflow
The pkgcheck action's container step runs as a different user than the
checkout, so libgit2 refused to open the repository and no checks ran.
Sets safe.directory before the composite action executes.
