+++ .github/workflows/pkgcheck.yaml
@@ -14,4 +14,7 @@ jobs:
   check:
     runs-on: ubuntu-latest
     steps:
+      - name: Mark workspace safe for git
+        run: git config --global --add safe.directory /github/workspace
+
       - uses: ropensci-review-tools/pkgcheck-action@main
