+++ reference/brapi_search_calls.md
@@ -38,7 +38,7 @@ A tibble of matching genotype calls.
 # \donttest{
 con <- brapi_connection("https://test-server.brapi.org")
 brapi_search_calls(con, variantSetDbIds = "variantset1")
-#> ℹ Async search started (ID: 2d2f82c5-bd59-4431-8f36-ef5040bbc685). Polling...
+#> ℹ Async search started (ID: ad164728-0186-404a-8722-23ef7d811f58). Polling...
 #> # A tibble: 260 × 12
 #>    additionalInfo callSetDbId callSetName genotype         genotypeValue
 #>    <lgl>          <chr>       <chr>       <list>           <chr>        
