+++ reference/brapi_cache_clear.md
@@ -25,9 +25,9 @@ Invisibly returns `con`.
 ``` r
 con <- brapi_connection("https://test-server.brapi.org")
 con <- brapi_cache_enable(con, dir = tempdir())
-#> ✔ Caching enabled at /tmp/Rtmp0ZXtYR (TTL: 3600s)
+#> ✔ Caching enabled at /tmp/Rtmp8U6l1W (TTL: 3600s)
 brapi_cache_clear(con)
-#> Warning: cannot remove file '/tmp/Rtmp0ZXtYR/bslib-45406f9f66735000132499f277df5132', reason 'Directory not empty'
-#> Warning: cannot remove file '/tmp/Rtmp0ZXtYR/downlit', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/Rtmp8U6l1W/bslib-45406f9f66735000132499f277df5132', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/Rtmp8U6l1W/downlit', reason 'Directory not empty'
 #> ✔ Cleared 6 cached response(s).
 ```
