### 2026-08-10 f47c78d
feat: add Genome Maps entity, fix brapi_get_marker_map() resource
brapi_get_marker_map() called /variants and returned reference-assembly
coordinates, which the test server leaves null and the vignette wrongly
attributed to the server. The spec makes Variant.start optional, and
marker positions live on /markerpositions, part of the Genome Maps
entity the package did not wrap.

These are different coordinate systems: Variant.start places a variant
on a reference assembly, MarkerPosition.position places a marker on a
named map in cM or bp, and one marker may sit on several maps.

Adds brapi_maps
[…]
