+++ man/brapi_maps.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/genome_maps.R
+\name{brapi_maps}
+\alias{brapi_maps}
+\title{List Genome Maps}
+\usage{
+brapi_maps(con, ...)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{...}{Additional query parameters.}
+}
+\value{
+A tibble with one row per genome map.
+}
+\description{
+List Genome Maps
+}
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+brapi_maps(con)
+}
+
+}
