+++ man/brapi_search_marker_positions.Rd
@@ -1,0 +1,48 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/genome_maps.R
+\name{brapi_search_marker_positions}
+\alias{brapi_search_marker_positions}
+\title{Search Marker Positions}
+\usage{
+brapi_search_marker_positions(
+  con,
+  mapDbIds = NULL,
+  variantDbIds = NULL,
+  linkageGroupNames = NULL,
+  minPosition = NULL,
+  maxPosition = NULL,
+  ...
+)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{mapDbIds}{Character vector. Filter by genome map IDs.}
+
+\item{variantDbIds}{Character vector. Filter by marker/variant IDs.}
+
+\item{linkageGroupNames}{Character vector. Filter by linkage group
+names.}
+
+\item{minPosition}{Integer. Minimum position, inclusive.}
+
+\item{maxPosition}{Integer. Maximum position, inclusive.}
+
+\item{...}{Additional search body parameters.}
+}
+\value{
+A tibble of matching marker positions.
+}
+\description{
+The \verb{/markerpositions} GET filter (see \code{\link[=brapi_marker_positions]{brapi_marker_positions()}})
+takes a single \code{variantDbId}; this search endpoint accepts many IDs at
+once, which is what \code{\link[=brapi_get_marker_map]{brapi_get_marker_map()}} uses internally when
+looking up positions for an entire variant set.
+}
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+brapi_search_marker_positions(con, variantDbIds = c("variant01", "variant02"))
+}
+
+}

[…]
