### 2026-08-10 28f68b2
docs: correct coverage claims, add BrAPI.R to overlap
DESCRIPTION claimed access to every endpoint; README and the vignette
comparison claimed full specification coverage. Parsing the OpenAPI YAML
at tag V2.1 gives 201 endpoints, 138 of them retrieval, of which the
package covers 56 across 32 of 36 entities. All claims now quantified.

Adds BrAPI.R (TriticeaeToolbox) to the overlap discussion as an httr
transport layer rather than an endpoint client, and drops the BrAPI
version comparison row, which BrAPI.R's version argument makes stale.

Also fixes README examples still calling brapi_get_mar
[…]
