+++ tests/testthat/test-integration.R
@@ -133,6 +133,46 @@ test_that("brapi_germplasm_pedigree returns a single-row tibble", {
   expect_true("germplasmDbId" %in% names(ped))
 })
 
+test_that("brapi_pedigree returns pedigree nodes with tidied relative columns", {
+  skip_on_cran()
+  skip_if_offline_brapi()
+
+  con <- brapi_connection(SERVER)
+  ped <- brapi_pedigree(
+    con,
+    includeParents = TRUE, includeSiblings = TRUE, includeProgeny = TRUE
+  )
+
+  skip_if(nrow(ped) == 0L, "No pedigree records on test server")
+
+  expect_s3_class(ped, "data.frame")
+  expect_true("germplasmDbId" %in% names(ped))
+  # As of this writing the server returns 3 pedigree nodes, at least one of
+  # which (germplasm3) has 2 populated parents - confirms the parents
+  # list-column really is tidied into per-node tibbles against live data,
+  # not just against the mocked shape.
+  expect_true("parents" %in% names(ped))
+  non_empty_parents <- vapply(ped$parents, nrow, integer(1)) > 0L
+  skip_if(!any(non_empty_parents), "No node with populated parents on test server")
+  one_node <- ped$parents[[which(non_empty_parents)[1L]]]
+  expect_s3_class(one_node, "data.frame")
+  expect_true(all(c("germplasmDbId", "germplasmName", "parentType") %in% names(one_node)))
+  expect_gt(nrow(one_node), 0L)
+})
+
+test_that("brapi_search_pedigree returns the same node shape as brapi_pedigree", {
+  skip_on_cran()
+  skip_if_offline_brapi()
+
+  con <- brapi_connection(SERVER)
+  ped <- brapi_search_pedigree(con, includeParents = TRUE)
+
+  skip_if(nrow(ped) == 0L, "No pedigree records on test server")
+
+  expect_s3_class(ped, "data.frame")
+  expect_true(all(c("germplasmDbId", "parents") %in% names(ped)))
+})
+
 # ---------------------------------------------------------------------------
 # Phenotyping module
 # ---------------------------------------------------------------------------
@@ -160,6 +200,53 @@ test_that("brapi_observation_units returns a non-empty tibble", {
   expect_s3_class(ou, "data.frame")
   expect_gt(nrow(ou), 0L)
   expect_true("observationUnitDbId" %in% names(ou))
+})
+
+test_that("brapi_ontologies returns a non-empty tibble", {
+  skip_on_cran()
+  skip_if_offline_brapi()
+
+  con <- brapi_connection(SERVER)
+  onto <- brapi_ontologies(con)
+
+  expect_s3_class(onto, "data.frame")
+  expect_gt(nrow(onto), 0L)
+  expect_true(all(c("ontologyDbId", "ontologyName") %in% names(onto)))
+})
+
+test_that("brapi_ontology returns a single-row tibble for a retrievable ID", {
+  skip_on_cran()
+  skip_if_offline_brapi()
+
+  con <- brapi_connection(SERVER)
+  onto <- brapi_ontologies(con)
+  skip_if(nrow(onto) == 0L, "No ontologies on test server")
+
+  # As of this writing, most ontologyDbIds returned by GET /ontologies
+  # (UUID-shaped ones) 404 on GET /ontologies/{id} - a real inconsistency
+  # in this server's seed data, confirmed directly with curl, not a
+  # brapiR2 bug. Only the small number of "O_NNN"-style IDs resolve. Try
[…]
