+++ tests/testthat/test-genotyping-mocked.R
@@ -105,65 +105,149 @@ test_that("brapi_get_dosage_matrix errors when no allele matrix is returned", {
   expect_error(brapi_get_dosage_matrix(con, "vs1"), "No allele matrix data")
 })
 
-test_that("brapi_get_marker_map returns tidy marker positions", {
+## brapi_get_marker_map() sources positions from the Genome Maps entity
+## (brapi_marker_positions()/brapi_search_marker_positions()), not from
+## brapi_variants() - it just uses brapi_variants() to look up which
+## variant IDs belong to a set, in the variantSetDbId path.
+
+MARKER_MAP_COLS <- c(
+  "variantDbId", "variantName", "mapDbId", "mapName", "type", "unit",
+  "linkageGroupName", "position"
+)
+
+test_that("brapi_get_marker_map(mapDbId=) joins map metadata onto positions", {
   local_mocked_bindings(
-    brapi_variants = function(con, variantSetDbId = NULL, ...) {
+    brapi_marker_positions = function(con, mapDbId = NULL, ...) {
       tibble::tibble(
-        variantDbId   = c("v1", "v2"),
-        variantName   = c("m1", "m2"),
-        referenceName = c("1", "1"),
-        start         = c(100L, 200L)
+        variantDbId      = c("v1", "v2"),
+        variantName      = c("m1", "m2"),
+        mapDbId           = c("map1", "map1"),
+        mapName           = c("Map One", "Map One"),
+        linkageGroupName  = c("Chromosome 1", "Chromosome 1"),
+        position          = c(100L, 200L)
+      )
+    },
+    brapi_maps = function(con, ...) {
+      tibble::tibble(
+        mapDbId = "map1", mapName = "Map One", type = "Genetic", unit = "cM"
       )
     },
     .package = "brapiR2"
   )
 
   con <- brapi_connection("https://example.org")
-  markers <- brapi_get_marker_map(con, "vs1")
+  markers <- brapi_get_marker_map(con, mapDbId = "map1")
 
   expect_identical(nrow(markers), 2L)
-  expect_identical(
-    names(markers),
-    c("variantDbId", "variantName", "referenceName", "start")
-  )
-  expect_identical(markers$variantDbId, c("v1", "v2"))
-  expect_identical(markers$variantName, c("m1", "m2"))
-  expect_identical(markers$referenceName, c("1", "1"))
-  expect_identical(markers$start, c(100L, 200L))
+  expect_identical(names(markers), MARKER_MAP_COLS)
+  expect_identical(markers$type, c("Genetic", "Genetic"))
+  expect_identical(markers$unit, c("cM", "cM"))
+  expect_identical(markers$position, c(100L, 200L))
 })
 
-test_that("brapi_get_marker_map handles a variantNames list-column", {
+test_that("brapi_get_marker_map(variantSetDbId=) bridges through search", {
+  captured <- new.env()
   local_mocked_bindings(
     brapi_variants = function(con, variantSetDbId = NULL, ...) {
+      tibble::tibble(variantDbId = c("v1", "v2"))
+    },
+    brapi_search_marker_positions = function(con, variantDbIds = NULL, ...) {
+      captured$variantDbIds <- variantDbIds
       tibble::tibble(
-        variantDbId   = "v1",
-        variantNames  = list(list("m1")),
-        referenceName = "1",
-        start         = 100L
+        variantDbId      = c("v1", "v2"),
+        variantName      = c("m1", "m2"),
+        mapDbId           = c("map1", "map1"),
+        linkageGroupName  = c("Chromosome 1", "Chromosome 1"),
+        position          = c(100L, 200L)
+      )
+    },
[…]
