### 2026-08-10 2948627
feat: add Pedigree and Ontologies entities
brapi_pedigree() and brapi_search_pedigree() make batch pedigree
retrieval a single call rather than one per germplasm. Nodes return one
row each with parents, siblings and progeny as tidied list-columns,
since the three relation types do not share a row schema and BrAPI's
model is node-centric.

brapi_ontologies() and brapi_ontology() cover the ontologies that
traits, scales, methods and variables reference.
