+++ R/germplasm.R
@@ -1,5 +1,6 @@
 # ---- BrAPI Germplasm Module ----
-# Endpoints: /germplasm, /attributes, /crosses, /crossingprojects, /seedlots
+# Endpoints: /germplasm, /attributes, /crosses, /crossingprojects,
+#            /seedlots, /pedigree, /search/pedigree
 
 
 #' List Germplasm
@@ -182,4 +183,183 @@ brapi_search_germplasm <- function(con,
     ...
   ))
   brapi_post_search(con, "/search/germplasm", body = body)
+}
+
+
+#' List Pedigree Nodes
+#'
+#' Retrieves a filtered subset of a pedigree tree via `/pedigree` - a batch
+#' endpoint for pulling pedigree records across many germplasm in one call.
+#' This is different from [brapi_germplasm_pedigree()], which retrieves one
+#' germplasm's immediate pedigree via the germplasm sub-resource
+#' (`/germplasm/{germplasmDbId}/pedigree`) and must be called once per
+#' germplasm. Use `brapi_pedigree()` (or [brapi_search_pedigree()]) to pull
+#' pedigree records for many germplasm at once - e.g. everything in a crop,
+#' program, or family - in one or a few requests; use
+#' [brapi_germplasm_pedigree()] when you already have a single germplasm ID
+#' in hand.
+#'
+#' Each row is one pedigree node (one germplasm). The server only includes
+#' a node's relatives if asked: set `includeParents`, `includeSiblings`,
+#' and/or `includeProgeny` to `TRUE` to populate the `parents`, `siblings`,
+#' and `progeny` list-columns, each holding a small tibble of related
+#' germplasm (`germplasmDbId`, `germplasmName`, and `parentType` - `NA` for
+#' siblings, which have none) that you can [tidyr::unnest()] when you need
+#' one row per relationship rather than one row per node. Nodes are never
+#' collapsed or flattened by default: a pedigree is graph-shaped (each node
+#' has its own parents, siblings, and progeny edges), and the three
+#' relation types don't share a common row shape, so there is no lossless
+#' single flat table to fall back to.
+#'
+#' @inheritParams brapi_shared_params
+#' @param germplasmDbId Character or NULL. Filter by germplasm.
+#' @param includeParents Logical or NULL. Include each node's parents.
+#' @param includeSiblings Logical or NULL. Include each node's siblings.
+#' @param includeProgeny Logical or NULL. Include each node's progeny.
+#' @param includeFullTree Logical or NULL. Recursively include every node
+#'   reachable in the pedigree tree.
+#' @param pedigreeDepth Integer or NULL. Number of levels to include up
+#'   the tree (parents, grandparents, ...).
+#' @param progenyDepth Integer or NULL. Number of levels to include down
+#'   the tree (children, grandchildren, ...).
+#'
+#' @return A tibble with one row per pedigree node. `parents`, `siblings`,
+#'   and `progeny`, when requested, are list-columns of small tibbles (one
+#'   row per relative) rather than raw nested lists or a flattened table.
+#'
+#' @seealso [brapi_germplasm_pedigree()] for one germplasm's pedigree via
+#'   the germplasm sub-resource; [brapi_search_pedigree()] for the same
+#'   batch retrieval via `POST`, with a fuller set of filters.
+#'
+#' @examples
+#' \donttest{
+#' con <- brapi_connection("https://test-server.brapi.org")
+#' brapi_pedigree(con, includeParents = TRUE, includeProgeny = TRUE)
+#' }
+#'
+#' @export
+brapi_pedigree <- function(con, germplasmDbId = NULL, includeParents = NULL,
[…]
