+++ man/brapi_ontologies.Rd
@@ -1,0 +1,36 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/phenotyping.R
+\name{brapi_ontologies}
+\alias{brapi_ontologies}
+\title{List Ontologies}
+\usage{
+brapi_ontologies(con, ...)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{...}{Additional query parameters.}
+}
+\value{
+A tibble with one row per ontology.
+}
+\description{
+Retrieves the ontologies registered on the server: metadata about each
+ontology (name, version, authors, description, ...), not the trait
+terms that belong to it. \code{\link[=brapi_traits]{brapi_traits()}}, \code{\link[=brapi_scales]{brapi_scales()}},
+\code{\link[=brapi_methods]{brapi_methods()}}, and \code{\link[=brapi_observation_variables]{brapi_observation_variables()}} each carry an
+ontology reference back to one of these records.
+}
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+brapi_ontologies(con)
+}
+
+}
+\seealso{
+\code{\link[=brapi_ontology]{brapi_ontology()}} for a single ontology by ID;
+\code{\link[=brapi_traits]{brapi_traits()}}, \code{\link[=brapi_scales]{brapi_scales()}}, \code{\link[=brapi_methods]{brapi_methods()}}, and
+\code{\link[=brapi_observation_variables]{brapi_observation_variables()}} for the records that reference these
+ontologies.
+}
