+++ man/brapi_ontology.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/phenotyping.R
+\name{brapi_ontology}
+\alias{brapi_ontology}
+\title{Get a Single Ontology by ID}
+\usage{
+brapi_ontology(con, ontologyDbId)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{ontologyDbId}{Character. The unique ontology identifier.}
+}
+\value{
+A single-row tibble with ontology details.
+}
+\description{
+Get a Single Ontology by ID
+}
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+brapi_ontology(con, "O_001")
+}
+
+}
+\seealso{
+\code{\link[=brapi_ontologies]{brapi_ontologies()}}; \code{\link[=brapi_traits]{brapi_traits()}}, \code{\link[=brapi_scales]{brapi_scales()}},
+\code{\link[=brapi_methods]{brapi_methods()}}, and \code{\link[=brapi_observation_variables]{brapi_observation_variables()}} for the
+records that reference ontologies.
+}

[…]
