+++ reference/brapi_methods.md
@@ -24,6 +24,14 @@ brapi_methods(con, ...)
 
 A tibble with one row per measurement method.
 
+## See also
+
+[`brapi_ontologies()`](https://josh45-source.github.io/brapiR2/reference/brapi_ontologies.md)
+and
+[`brapi_ontology()`](https://josh45-source.github.io/brapiR2/reference/brapi_ontology.md)
+to resolve the ontology a method's `ontologyDbId`/`ontologyReference`
+points to.
+
 ## Examples
 
 ``` r

[…]
