### 2026-08-10 5bae452
docs: note test server map unit inconsistency, balance QBMS comparison
The test server's genome maps declare type Physical Map with unit cM,
which contradicts the sentence above the output explaining that units
follow the map's type. Notes this as fixture data rather than package
or specification behaviour.

The comparison table gave brapiR2 a counted coverage figure and QBMS an
uncounted characterisation; the QBMS cell now points to their own
documentation. Adds BrAPI.R to Related Packages, where it was discussed
in prose but missing from the list.
