+++ reference/brapi_search_calls.md
@@ -38,7 +38,7 @@ A tibble of matching genotype calls.
 # \donttest{
 con <- brapi_connection("https://test-server.brapi.org")
 brapi_search_calls(con, variantSetDbIds = "variantset1")
-#> ℹ Async search started (ID: e9a5fd80-165b-472e-ad9c-c0d31e7b381c). Polling...
+#> ℹ Async search started (ID: dd6eac7e-8a2e-4106-b62a-93d9efb75a09). Polling...
 #> # A tibble: 260 × 12
 #>    additionalInfo callSetDbId callSetName genotype         genotypeValue
 #>    <lgl>          <chr>       <chr>       <list>           <chr>        
