+++ articles/genomic-selection-pipeline.md
@@ -99,7 +99,7 @@ vs_id <- vsets$variantSetDbId[1]
 ## each marker on a named map (genetic, cM, or physical, bp) and is
 ## populated here.
 markers <- brapi_get_marker_map(con, variantSetDbId = vs_id)
-#> ℹ Async search started (ID: e28c1d0c-ff05-451e-bbf6-a0293b6b5f13). Polling...
+#> ℹ Async search started (ID: 26705aa0-21a5-480b-b70f-05dac94dfd89). Polling...
 #> Warning: 14 of 20 variants in "variantset1" have no marker position record; returning
 #> positions for the remaining 6.
 markers
