### 2026-08-17 #792 @robitalec
Thanks to @jmh579 and @dwaring87 for agreeing to review.
Please see the [Guide for Reviewers](https://devguide.ropensci.org/softwarereview_reviewer.html) and let me know when you have any questions.

@josh45-source, I've reviewed your changes regarding my initial editor comments and I'll ask you now to limit any major changes to the package while the reviewers are working on their reviews.

Thanks everyone!
