### 2026-08-21 #792 @robitalec
Just noticed - gentle reminder @josh45-source to add the rOpenSci review status badge to the package README. See [above](https://github.com/ropensci/software-review/issues/792#issuecomment-5284053170). Thanks!
