### 2026-08-25 41dd317
docs: add rOpenSci review badge, note map unit inconsistency, balance QBMS comparison
Adds the software peer review status badge for issue 792, as requested
by the review bot.

The test server's genome maps declare type Physical Map with unit cM,
which contradicts the sentence above the output explaining that units
follow the map's type. Notes this as fixture data rather than package
or specification behaviour.

The comparison table gave brapiR2 a counted coverage figure and QBMS an
uncounted characterisation; the QBMS cell now points to their own
documentation. Adds BrAPI.R to Related Package
[…]
