+++ reference/brapi_seed_lots.md
@@ -34,7 +34,7 @@ brapi_seed_lots(con)
 #>   additionalInfo externalReferences amount createdDate germplasmDbId lastUpdated
 #>   <list>         <list>              <dbl> <chr>       <chr>         <chr>      
 #> 1 <named list>   <list [1]>            360 2020-04-02… germplasm1    2020-04-08…
-#> 2 <named list>   <list [1]>             40 2020-04-02… germplasm3    2020-04-08…
+#> 2 <named list>   <list [1]>             40 2020-04-02… germplasm2    2020-04-08…
 #> # ℹ 11 more variables: locationDbId <chr>, locationName <chr>,
 #> #   programDbId <chr>, programName <chr>, seedLotDescription <chr>,
 #> #   seedLotName <chr>, sourceCollection <chr>, storageLocation <chr>,
