+++ reference/brapi_cache_clear.md
@@ -25,9 +25,9 @@ Invisibly returns `con`.
 ``` r
 con <- brapi_connection("https://test-server.brapi.org")
 con <- brapi_cache_enable(con, dir = tempdir())
-#> ✔ Caching enabled at /tmp/Rtmp9WG7Xr (TTL: 3600s)
+#> ✔ Caching enabled at /tmp/RtmprtUbmh (TTL: 3600s)
 brapi_cache_clear(con)
-#> Warning: cannot remove file '/tmp/Rtmp9WG7Xr/bslib-45406f9f66735000132499f277df5132', reason 'Directory not empty'
-#> Warning: cannot remove file '/tmp/Rtmp9WG7Xr/downlit', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/RtmprtUbmh/bslib-e9b2b13fa612f50d23e4850d93d60d01', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/RtmprtUbmh/downlit', reason 'Directory not empty'
 #> ✔ Cleared 6 cached response(s).
 ```
