### 2026-08-25 #792 @josh45-source
apologies @robitalec  for the delay, I have added the badge, https://github.com/josh45-source/brapiR2/commit/41dd317
