### 2026-09-02 #792 @dwaring87
- I have no working relationship with the package author.

- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

    - As noted in the package submission, I have developed a similar R package for aiding in BrAPI requests.  I agree with Joash's statement that the packages are more complimentary than competing and serve different use-cases.

	- This could be user-error on my end, but I'm not sure how to view the vignettes in this package.  I see that they are present in the GitHub repo, but calling `vignette(package="brapiR2")` yields a `no vignettes found` error.

	- It may be worth setting the user-agent on requests made by the package so the source can be identified on the server
	- It could be useful to expose and document some of the internal request functions to allow a user to make their own BrAPI request to an endpoint that is not supported by a function in this package.
	- The LICENSE is not open source

---

Overall, I think this is a very useful package.  It aims to be a user-friendly interface to extracting data from any BrAPI-compliant database.  It can be used by those that are not familiar with the BrAPI specification or the specific endpoints they would need to use to fetch their desired data.  The fact that all of the functions are pipeable and are standardized to return data in tibbles makes the package a very useful component in data analysis pipelines.  Thanks for taking the time to build this and making it available for others!

#### Issues
I tested the package with the BrAPI Test Server and the functionality worked well with the simple data on the test server.  However, I ran into a few issues testing some of the functions with real (and somewhat messy, but realistic) data from a Breedbase database (specifically [T3/Wheat](https://wheat.triticeaetoolbox.org)).

- **Login with username / password did not work with breedbase:**

The package README has this snippet about logging in to a breedbase instance:

```
# Token-based (most BrAPI servers)
con <- brapi_connection("https://my-breedbase.org")
con <- brapi_login(con, "username", "password")
```

However, when I tried this with a few different breedbase instances, the login would fail with an error message.

```R
> cb <- brapi_connection("https://cassavabase.org")
> cb <- brapi_login(cb, "my_username", "my_password")
Error in `req_perform()`:
! HTTP 400 Bad Request.
Run `rlang::last_trace()` to see where the error occurred.
```

I looked into this a bit more and discovered this is **not a problem with the brapiR2 package**, but a problem with how breedbase is handling the request.  This should be fixed soon in a future update to breedbase ([PR #6272](https://github.com/solgenomics/sgn/pull/6272)).

- **Missing functions:**

Overall, the package includes great coverage for almost all of the BrAPI endpoints that would be used for fetching data from a database.  The naming convention used for the functions is very straight-forward.  There is generally a `brapi_{item}s(con, ...)` function for fetching all of the items of a particular data type and a `brapi_{item}(con, id, ...)` function for fetching a specific single item of a data type.

There are a few instances where there is a missing function for getting a single item by id.  For example, there is a `brapi_locations(con)` function but no corresponding `brapi_location(con, id)` function.  A possible use-case for the missing function would be a user has a study they are working with and they know the id of the location for that study.  They then want to get the geo coordinates for that location.  Currently, they would have to get all of the locations and filter the result by id.  A slightly easier approach would be to use a function to fetch the data for just that location.

Another example is with lists.  There is a `brapi_lists(con)` function which returns metadata about the user's lists (but not the contents of the list).  However, there is no corresponding `brapi_list(con, id)` function which is required to get the actual contents of the list.

- **get_study_data:**

This function seems like a very useful helper tool to get data from different BrAPI endpoints into a more useful format for a very common use-case.  I tested it with a few different trials and it worked for most of them, but it did fail with some edge cases.  For the ones that failed, it seemed to fail with the `mutate` stage where it's trying to simplify lists

> Error in `dplyr::mutate()`:
> ℹ In argument: `dplyr::across(...)`.
> Caused by error in `across()`:
> ! Can't compute column `Pea plant height - cm|COMP:0000010`.
> Caused by error in `dplyr_internal_error()`:
> Run `rlang::last_trace()` to see where the error occurred.

In this example, the Pea plant height trait had a lot of missing values, where the values are set to `NULL` for a lot of the plots.  Here is what the `wide` data looks like before the call to `mutate`:

I can give you more details if you want examples to debug.

- **Deprecated Endpoints:**

The `brapi_germplasm_pedigree()` function calls the `/germplasm/{germplasmDbId}/pedigree` BrAPI endpoint.  This endpoint is marked as deprecated as of v2.1 in the BrAPI specification and should be replaced with `/pedigree?germplasmDbId={germplasmDbId}`.

The `brapi_germplasm_progeny()` function uses a similarly deprecated BrAPI endpoint.

#### Suggestions

- **Error Reporting**

When an endpoint would fail (for example, the login request to a breedbase instance), the returned error message would just display the generic HTTP status code and message.  In many cases, the server would return a more specific and useful message in the body of the response (such as the `metadata.status[].message` field).

```json
{
  "result": null,
  "metadata": {
    "pagination": {
      "totalCount": 0,
      "currentPage": 0,
      "pageSize": 1,
      "totalPages": 0
    },
    "datafiles": [],
    "status": [
      {
        "messageType": "ERROR",
        "message": "JSON array body required"
      }
    ]
  }
}
```

Displaying these more specific error messages would make it much easier for a user to debug a problem.

- **Documented Parameters**

It is great that the package's functions support passing any/all arguments as parameters to the underlying BrAPI endpoints.  This allows the functions to be very flexible to make custom queries and allows support for databases that may not strictly follow the BrAPI specification.  However, the documentation in the package is a bit sparse on what parameters are possible.  If you're not familiar with the BrAPI specification, you wouldn't know which parameters are supported.

For example, the `brapi_studies(con, trialDbId = NULL, ...)` function is documented with just the `trialDbId` parameter listed.  Most BrAPI databases will also support many other parameters that are included in the BrAPI specification (such as `locationDbId`, `seasonDbId`, `programDbId`, etc).

My suggestion would be to either include the parameters that are listed in the BrAPI specification for each endpoint in the R documentation for each function or to list the BrAPI endpoint that is used for the function with a link to the BrAPI specification.  This would allow the users to know which parameters can be included when using each function.

- **brapi_study_data() fallback**

In debugging the issue with the `brapi_study_data()` function, I noticed the fallback functionality that was added to handle the case when fetching the observation data for a single study does not return any data.  In this case, the function will make a second request to fetch all observations and filter the data client-side for the requested study.  This seems like a useful way to handle the case when the BrAPI server does not return any data.

It may be useful to add another check to make sure the returned data is indeed only from the requested study.  There may be some servers that don't read the `studyDbId` parameter and may unexpectedly return all observations.  In this case, if there is more than one `studyDbId` in the returned data, the function could simply filter the data to only include observations from the requested `studyDbId`.
