### 2026-09-02 #792 @robitalec
Thanks @dwaring87 for the excellent review. I really appreciate your deep understanding of the Breeding API specifications, endpoints and also where {brapiR2} is especially helpful from the user's experience.

@josh45-source, I'll remind you to please hold off on any major changes while our second reviewer is working.
However, please feel free to have any discussions, ask for any clarifications, etc with @dwaring87 regarding their review in the meantime.

Thanks everyone!
