### 2026-09-03 #792 @josh45-source
Thank you for this, @dwaring87 , Testing against T3/Wheat and Cassavabase rather than the reference server is what makes the review useful, and it found things I could not have found on my own. The public test server has three studies, two of which hold no observations at all, so almost nothing in your report could have surfaced against it.

Following @robitalec  note I am holding off on changes until the second review lands. Below is what I understand so far, and a few questions where your view would be of great help.

The licence is a real gap, and thank you for catching it. `LICENSE` contains only the year and copyright holder, which is the companion file `usethis::use_mit_license()` writes alongside `LICENSE.md`. I have the companion without the licence text, so anyone opening that file finds no grant of rights at all. `usethis::use_mit_license()` will write both properly. The copyright holder also reads "brapiR2 authors", which should be my name.

The vignette question is my documentation's fault rather than yours. `install_github()` does not build vignettes unless asked, so `install_github("josh45-source/brapiR2", build_vignettes = TRUE)` is needed. The README installation instructions do not say so and will.

If you want to confirm on your side, `tools::getVignetteInfo("brapiR2")` after a build-vignettes install should return one row for `brapiR2.Rmd`. Worth knowing that `vignette()` prints through a pager, so it looks silent under `Rscript` even when the vignette is correctly installed, which cost me a few minutes checking this just now.

One clarification on scope: only `brapiR2.Rmd` is a vignette. The genomic selection and multi-environment documents are pkgdown articles, `.Rbuildignore`d because they depend on rrBLUP, BGLR, sommer, metan and lme4, which would otherwise become build-time dependencies. They are on the site but never in the installed package.

`brapi_study_data()`. Your screenshot is diagnostic, thank you. `Plant height - cm|CO_350:0000232` is a list-column of three-element character vectors in every row, and simplifies without trouble. `Pea plant height - cm|COMP:0000010` is mixed: `NULL` in most rows, bare scalars like `50`, `125` and `132` in others. Both columns carry equally non-syntactic names, spaces and a pipe, and only the second one fails, so the name is not the cause after all. My simplification step assumes a consistent shape down each column and cannot assemble a result when some elements are `NULL` and others are length-one atomics, which fits the `dplyr_internal_error()` in the traceback.

Trial IDs would still help me reproduce it directly rather than reason from the screenshot, if you have them to hand.

One thing I noticed while reading it that you did not report: the working column comes back as `c("150", "80 ", "91 ")`, with trailing whitespace on the values. Not the bug you hit, but worth trimming on my side before anyone converts them to numeric.

On the fallback, your suggestion to verify the returned data really is from the requested study is well aimed, and there is history worth sharing. That check existed but was broken: the filter read `.data$studyDbId == studyDbId`, and since the function's own argument shares the column's name, dplyr's data mask resolved the bare symbol to the column. The comparison was always TRUE, so any requested study returned the first study's observations. Fixed a few weeks ago with a regression test that fails against the old behaviour. It surfaced only because my editor pushed me to evaluate vignette code that had been `eval = FALSE`. So your instinct that this path deserves scrutiny was right; it had already bitten once.

Deprecated endpoints. You are right, and it is slightly awkward on my side: I added `brapi_pedigree()` and `brapi_search_pedigree()` a few weeks ago and described the older function as the single-germplasm route rather than the deprecated one.

Here I would rather have your view than pick. Would you expect `brapi_germplasm_pedigree()` to quietly redirect to `/pedigree?germplasmDbId=` and keep working, to warn while redirecting, or to be formally deprecated in favour of the new functions? Same question for `brapi_germplasm_progeny()`. My instinct is a silent redirect, since the user asked for one germplasm's pedigree and should not have to care which endpoint serves it, but that hides a specification change they may want to know about.

Error reporting. Clearly right and I should have done it already. BrAPI puts a real message in `metadata.status[].message` and I discard it in favour of the HTTP status. Your example, "JSON array body required" surfacing as a bare 400, makes the case by itself. I will surface the status messages where present and fall back to the HTTP code where not.

Missing by-ID functions. Also right, and the location example is the convincing one: knowing a `locationDbId` from a study and having to fetch every location to get its coordinates is silly. I added several by-ID functions in this shape recently, and `brapi_location()` and `brapi_list()` belong with them. `brapi_list()` is the more serious gap, since without it there is no way to reach a list's contents at all.

Documented parameters. I would rather ask than choose. Listing every specification parameter per function is more useful at the point of reading but goes stale whenever BrAPI adds a field. Linking the endpoint stays correct indefinitely and is cheaper to maintain, since I have the v2.1 OpenAPI YAML parsed already and could generate those links mechanically. As a user of a BrAPI client, which would you rather find in a help page? Or both, with the common filters listed and a link for the full set?

Exposing the request internals. This one I would rather think about with you than around you, since it moves brapiR2 toward what BrAPI.R already does well. Someone needing an endpoint I have not wrapped can compose the call in BrAPI.R today and get a working result. Exposing `brapi_get()` and `brapi_post_search()` would give them a rawer path through my package instead, with pagination and async search handled but no typed function at the end of it. Do you see that as genuinely useful, or is the more honest answer to point users at BrAPI.R for that case and keep brapiR2 to the named-function contract? I can see it either way and you are better placed to judge.

On the user agent, agreed, straightforwardly. Servers should be able to see what is making the request.

One thing from my side that your review has bearing on. The DESCRIPTION claims the package works with BreedBase, BMS, EBS, GIGWA and Germinate, and until now the only server it had ever been tested against was `test-server.brapi.org`. You have given me the first real evidence for BreedBase, including a login failure that turned out to be theirs rather than mine. I intend to narrow that claim to what has actually been verified, and to test properly against the public T3 databases and Cassavabase. Your `listBrAPIConnections()` is how I learned those servers existed, so thank you for that too.

Thanks again for the six hours spent on this work. The findings that matter all came from you using real data, which is precisely the part I could not do myself.
