### 2026-09-04 #792 @dwaring87
@josh45-source Thanks for the prompt and detailed reply!  I'll respond inline below:

> The licence is a real gap, and thank you for catching it. LICENSE contains only the year and copyright holder, which is the companion file usethis::use_mit_license() writes alongside LICENSE.md. I have the companion without the licence text, so anyone opening that file finds no grant of rights at all. usethis::use_mit_license() will write both properly. The copyright holder also reads "brapiR2 authors", which should be my name.

I now see that MIT is referenced in the DESCRIPTION file, but it would be good to have the full text of the license in the separate LICENSE file (I don't see LICENSE.md anywhere).  GitHub automatically links to the LICENSE file in the sidebar of the repo and in the tab with the README, so it would be good to have it fully included there.

> The vignette question is my documentation's fault rather than yours. install_github() does not build vignettes unless asked, so install_github("josh45-source/brapiR2", build_vignettes = TRUE) is needed. The README installation instructions do not say so and will.

Thanks for noting the `build_vignettes = TRUE` argument, I wasn't aware the vignettes were opt-in when installing from source - that is good to know.  I reinstalled the package and the vignette now shows up.  I think the vignette and the articles are very thorough and are great examples of what you can do with the package and highlight some of the most common use-cases.

> My simplification step assumes a consistent shape down each column and cannot assemble a result when some elements are NULL and others are length-one atomics, which fits the dplyr_internal_error() in the traceback.

I think this is important to fix.  Real breeding data is often very messy and it is common to have missing values with trait observations.

> Trial IDs would still help me reproduce it directly rather than reason from the screenshot, if you have them to hand.

The example I gave here was a simplified test trial on a development server, but studyDbId 6882 on https://oat-sandbox.triticeaetoolbox.org is another example that fails - it also has some traits with missing data.

> One thing I noticed while reading it that you did not report: the working column comes back as c("150", "80 ", "91 "), with trailing whitespace on the values. Not the bug you hit, but worth trimming on my side before anyone converts them to numeric.

I had not noticed the trailing whitespace - that is a good catch.  Ideally, that is something that we should fix on the database.  But I think it would be a good idea for the package to do some simple cleanup since that may happen in other instances.

> Here I would rather have your view than pick. Would you expect brapi_germplasm_pedigree() to quietly redirect to /pedigree?germplasmDbId= and keep working, to warn while redirecting, or to be formally deprecated in favour of the new functions? Same question for brapi_germplasm_progeny(). My instinct is a silent redirect, since the user asked for one germplasm's pedigree and should not have to care which endpoint serves it, but that hides a specification change they may want to know about.

I agree that a slient redirect should be fine.  I think the nice thing about your package is that it abstracts away the specifics about the BrAPI specification so user's don't need to be bothered by the API endpoints.  That way you can keep the package functions working as the BrAPI spec changes over time.

> As a user of a BrAPI client, which would you rather find in a help page? Or both, with the common filters listed and a link for the full set?

This is a hard design choose.  As an R user, I would likely prefer to have all the documentation in the R function.  But that does add a maintenance cost to you to keep that documentation up to date.  If there is some way to automate that using the OpenAPI specification, that would be great, but that may be difficult to implement.  For now, I think it would be fine to just link to the BrAPI specification and think about a more elegant solution.

> Do you see that as genuinely useful, or is the more honest answer to point users at BrAPI.R for that case and keep brapiR2 to the named-function contract? I can see it either way and you are better placed to judge.

Thanks for recognizing BrAPI.R in your consideration.  Personally, I'm not concerned about the potential overlap.  I think if a user reaches for your package due to the ease of use of having named functions for most of the endpoints, it would be good to allow them the flexibility of making direct requests without needing to install another package.  It is also a suggestion in the [rOpenSci API client guidelines](https://devguide.ropensci.org/pkg_building.html#api-clients).
