### 2026-09-16 6ea50c4
Regenerate NAMESPACE under roxygen2 8.1.0
Multi-line importFrom formatting only; no change to exports or to which
functions are imported.
