+++ NAMESPACE
@@ -63,46 +63,62 @@ export(brapi_trials)
 export(brapi_variant_sets)
 export(brapi_variants)
 export(is_brapi_con)
-importFrom(cli,cli_abort)
-importFrom(cli,cli_alert_info)
-importFrom(cli,cli_alert_success)
-importFrom(cli,cli_alert_warning)
-importFrom(cli,cli_progress_bar)
-importFrom(cli,cli_progress_done)
-importFrom(cli,cli_progress_update)
-importFrom(cli,cli_warn)
-importFrom(dplyr,bind_rows)
-importFrom(dplyr,filter)
-importFrom(dplyr,mutate)
-importFrom(dplyr,select)
+importFrom(cli,
+  cli_abort,
+  cli_alert_info,
+  cli_alert_success,
+  cli_alert_warning,
+  cli_progress_bar,
+  cli_progress_done,
+  cli_progress_update,
+  cli_warn
+)
+importFrom(dplyr,
+  bind_rows,
+  filter,
+  mutate,
+  select
+)
 importFrom(glue,glue)
-importFrom(httr2,req_auth_bearer_token)
-importFrom(httr2,req_body_json)
-importFrom(httr2,req_headers)
-importFrom(httr2,req_method)
-importFrom(httr2,req_perform)
-importFrom(httr2,req_retry)
-importFrom(httr2,req_url_path_append)
-importFrom(httr2,req_url_query)
-importFrom(httr2,request)
-importFrom(httr2,resp_body_json)
-importFrom(httr2,resp_status)
-importFrom(jsonlite,fromJSON)
-importFrom(jsonlite,toJSON)
-importFrom(purrr,compact)
-importFrom(purrr,map)
-importFrom(purrr,map_chr)
-importFrom(purrr,map_dfr)
-importFrom(purrr,map_lgl)
-importFrom(rlang,"%||%")
-importFrom(rlang,.data)
-importFrom(rlang,.env)
-importFrom(rlang,abort)
-importFrom(rlang,hash)
-importFrom(rlang,inform)
-importFrom(rlang,warn)
-importFrom(tibble,as_tibble)
-importFrom(tibble,tibble)
-importFrom(tidyr,pivot_wider)
-importFrom(tidyr,unnest)
+importFrom(httr2,
+  req_auth_bearer_token,
+  req_body_json,
+  req_headers,
+  req_method,
+  req_perform,
+  req_retry,
+  req_url_path_append,
+  req_url_query,
+  request,
+  resp_body_json,
+  resp_status
+)
+importFrom(jsonlite,
+  fromJSON,
+  toJSON
+)
+importFrom(purrr,
+  compact,
+  map,
+  map_chr,
+  map_dfr,
+  map_lgl
+)
+importFrom(rlang,
+  "%||%",
+  .data,
+  .env,
+  abort,
+  hash,
+  inform,
+  warn
+)
+importFrom(tibble,
+  as_tibble,
+  tibble
+)
+importFrom(tidyr,
+  pivot_wider,
+  unnest
+)
 importFrom(tidyselect,where)
