### 2026-09-16 9bf2cfd
Add full MIT licence text and record reviewers
LICENSE held only the year and copyright holder, with no LICENSE.md
alongside it, so the repository contained no grant of rights and GitHub
had no licence to detect. Adds LICENSE.md and replaces 'brapiR2 authors'
with the named copyright holder in both files.

Adds the cph role, and David Waring and Jenna Hershberger with the rev
role, per their consent in the review thread.
