### 2026-09-16 72ff8fc
Documentation pass from reviewer feedback: licence, spelling, README
Records both reviewers with the rev role and restores the maintainer ORCID, which was dropped when Authors@R was rewritten.

Switches the README install instruction to pak::pak() and documents build_vignettes = TRUE for remotes and devtools. Corrects 'Genotypics' in the vignette comparison table. Declares en-GB and corrects six American spellings to match. Disables pandoc smart punctuation so README.md and README.Rmd stay in sync. Settles on 'Breedbase' throughout.
