+++ authors.md
@@ -3,6 +3,7 @@
 ## Authors
 
 - **Joash Joshua Ayo**. Author, maintainer, copyright holder.
+  [](https://orcid.org/0009-0007-1642-0172)
 
 - **David Waring**. Reviewer.  
   David reviewed the package (v. 0.1.0) for rOpenSci, see
