### 2026-09-16 45e53e6
Add a path argument to brapi_connection() for non-standard deployments
BrAPI was assumed to live at /brapi/ on every server. GRIN-Global serves it at /gringlobal/brapi/, so USDA-GRIN instances were unreachable; Germinate and GIGWA deployments commonly sit under their own prefixes too. The path is now a connection argument defaulting to 'brapi', so existing code is unaffected.

The cache key includes the path, so two servers sharing a hostname no longer collide, and print() shows it only when it differs from the default. Found while probing public servers for a compatibility report rather than 
[…]
