+++ articles/brapiR2.md
@@ -244,7 +244,7 @@ now reads from there instead:
 
 # Positions for every variant in the set, wherever they have been placed
 markers <- brapi_get_marker_map(con, variantSetDbId = vs_id)
-#> ℹ Async search started (ID: c4266f64-f116-4127-8257-df0ee8fc0dff). Polling...
+#> ℹ Async search started (ID: f542baf3-2a48-473b-ba3b-e6fa47401c74). Polling...
 #> Warning: 14 of 20 variants in "variantset1" have no marker position record; returning
 #> positions for the remaining 6.
 markers
@@ -401,7 +401,7 @@ fetching don’t need authentication).
 cache_dir <- tempfile("brapi_cache_")
 dir.create(cache_dir)
 perf_con <- brapi_cache_enable(con, ttl = 3600, dir = cache_dir)
-#> ✔ Caching enabled at /tmp/RtmppacEdT/brapi_cache_1f115a1fdf31 (TTL: 3600s)
+#> ✔ Caching enabled at /tmp/Rtmpv3CHtt/brapi_cache_1ee94611982f (TTL: 3600s)
 
 # First call: hits the server
 invisible(brapi_programs(perf_con))
