+++ news/index.md
@@ -68,7 +68,13 @@
   [`brapi_observation_variables()`](https://josh45-source.github.io/brapiR2/reference/brapi_observation_variables.md).
 - brapiR2 now wraps 32 of the 36 BrAPI v2.1 entities across all four
   modules (56 of 138 retrieval endpoints); see `DESIGN.md` for the full
-  coverage breakdown and which entities remain uncovered.
+  coverage breakdown and which entities remain uncovered. serve BrAPI
+  under `/brapi/`. GRIN-Global instances use `gringlobal/brapi` and were
+  previously unreachable; Germinate and GIGWA deployments commonly sit
+  under their own prefixes too. Defaults to `"brapi"`, so existing code
+  is unaffected. [`print()`](https://rdrr.io/r/base/print.html) shows
+  the path only when it differs from the default, and the cache key now
+  includes it, so two servers sharing a hostname no longer collide.
 
 #### Deprecated
 
@@ -82,7 +88,7 @@
   to fetch in parallel. The `.workers` argument is deprecated -
   supplying it now emits a warning and has no effect.
 
-  #### Documentation
+#### Documentation
 
 - Added `LICENSE.md` with the full MIT licence text, which was missing
   from the repository, so GitHub had no licence to detect and anyone
@@ -90,12 +96,9 @@
   now named explicitly rather than “brapiR2 authors”
   ([@dwaring87](https://github.com/dwaring87),
   ropensci/software-review#792).
-
-- [@dwaring87](https://github.com/dwaring87) and
-  [@jmh579](https://github.com/jmh579) are recorded in `DESCRIPTION`
-  with the `rev` role for their rOpenSci reviews
+- @dwaring87 and [@jmh579](https://github.com/jmh579) are recorded in
+  `DESCRIPTION` with the `rev` role for their rOpenSci reviews
   (ropensci/software-review#792).
-
 - Corrected a misspelling in the QBMS comparison table in the
   getting-started vignette ([@jmh579](https://github.com/jmh579),
   ropensci/software-review#792).
