+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   brapiR2: brapiR2.html
   articles/genomic-selection-pipeline: genomic-selection-pipeline.html
   articles/multi-environment-analysis: multi-environment-analysis.html
-last_built: 2026-09-16T12:08Z
+last_built: 2026-09-16T13:46Z
 urls:
   reference: https://josh45-source.github.io/brapiR2/reference
   article: https://josh45-source.github.io/brapiR2/articles
