### 2026-09-16 63836ce
Fix a duplicated NEWS entry and an indented heading
