+++ NEWS.md
@@ -50,6 +50,7 @@
 * brapiR2 now wraps 32 of the 36 BrAPI v2.1 entities across all four
   modules (56 of 138 retrieval endpoints); see `DESIGN.md` for the full
   coverage breakdown and which entities remain uncovered.
+* `brapi_connection()` gains a `path` argument for servers that do not
   serve BrAPI under `/brapi/`. GRIN-Global instances use
   `gringlobal/brapi` and were previously unreachable; Germinate and
   GIGWA deployments commonly sit under their own prefixes too. Defaults
