### 2026-09-17 aa3eaab
Vignette and documentation edits from reviewer feedback
Rewrites the vignette's Caching and Parallel Fetching section to explain what each feature does and when to use it, rather than describing the future package at length; the design rationale moves to DESIGN.md. Links Authentication from the connection section, and moves the QBMS recommendation to sit with the comparison rather than after the references.

Guards the parallel chunks on furrr and future being installed, so the vignette builds where suggested packages are absent. Declares Language: en-GB and makes the prose consistent with it.
