+++ vignettes/brapiR2.Rmd
@@ -57,6 +57,12 @@ con
 Every function takes `con` as its first argument. No global state is
 modified.
 
+Most real servers require credentials. See
+[Authentication](#authentication) below for username/password login,
+OAuth 2.0, and bearer tokens, and
+[Handling Credentials Safely](#credential-handling) for keeping them out
+of your scripts.
+
 ## Exploring Programs and Trials
 
 BrAPI organises breeding data as a nesting of four levels, and the calls
@@ -302,8 +308,16 @@ history, and out of anything you might `git commit` or share.
 
 ## Caching and Parallel Fetching
 
-These run against the public test server again (caching and parallel
-fetching don't need authentication).
+Two features for when you are fetching the same data repeatedly, or a lot
+of it at once. Both run against the public test server here; neither needs
+authentication.
+
+**Caching** stores each response on disk under a key built from the URL and
+query parameters, and returns it without a network call for as long as the
+TTL lasts. Reach for it when you are developing a script against a slow or
+distant server and re-running the same calls, or when a knitted report
+would otherwise hit the server once per render. It is off by default, per
+connection, and never shared between connections.
 
 ```{r performance, eval = server_up}
 # Enable caching — repeated calls within the TTL return instantly
@@ -321,17 +335,16 @@ brapi_programs(perf_con)
 brapi_cache_clear(perf_con)
 ```
 
-`brapi_fetch_parallel()` does not set a parallel backend itself - it runs
-against whatever `future` plan is already active, sequential by default.
-The future package's best-practices vignette is explicit that the choice
-of backend belongs to the caller, not to a package: a function that quietly
-sets and restores a plan on every call still mutates session-wide state you
-didn't ask it to touch, and can silently replace a plan you configured
-deliberately (a specific worker count, a cluster spanning several
-machines, callr workers for extra isolation). So it's your call to make,
-before this function runs:
+**Parallel fetching** runs one call across many identifiers at once, which
+helps when you are pulling data for dozens of studies or germplasm and each
+request is slow. `brapi_fetch_parallel()` uses whatever `future` plan is
+already active — sequential unless you say otherwise — so choosing the
+backend, and shutting it down afterwards, is yours to do. See
+[`future`'s own documentation](https://future.futureverse.org/) for the
+available backends, and DESIGN.md for why the package does not set one for
+you:
 
-```{r parallel, eval = server_up}
+```{r parallel, eval = server_up && requireNamespace("furrr", quietly = TRUE) && requireNamespace("future", quietly = TRUE)}
 future::plan(future::multisession, workers = 2)
 
 # Fetch study data from multiple studies in parallel (all studies on the
@@ -349,7 +362,7 @@ Since you set the plan, shutting the workers back down when you're done
 with them is your responsibility too, not something the package does for
 you on exit:
 
-```{r cleanup-parallel, eval = server_up}
+```{r cleanup-parallel, eval = server_up && requireNamespace("future", quietly = TRUE)}
 future::plan(future::sequential)
 ```
 
@@ -424,13 +437,13 @@ data <- brapi_programs(con) |>
   brapi_study_data(con = con)
 ```
 
+Choose QBMS for interactive exploration in the console. Choose brapiR2 for
+reproducible scripts, pipelines, or when you need genotypic data and
+caching. Many users use both.
+
 ## References
 
 - Selby, P., Abbeloos, R., Backlund, J.E., et al., & The BrAPI Consortium
   (2019). BrAPI — an application programming interface for plant breeding
   applications. *Bioinformatics*, 35(20), 4147–4155.
   <https://doi.org/10.1093/bioinformatics/btz190>
-
-Choose QBMS for interactive exploration in the console. Choose brapiR2 for
-reproducible scripts, pipelines, or when you need genotypic data and
-caching. Many users use both.
