### 2026-09-17 9596f1b
Reorder the README to lead with the package's own functionality
The README opened with a QBMS comparison table and three paragraphs on other BrAPI clients, so a reader met three competitor comparisons before learning what brapiR2 covers. Supported BrAPI Modules now comes first, the comparisons have moved into Related Packages under their own subheadings, and Authentication sits before the extended examples.
