+++ index.md
@@ -9,42 +9,25 @@ returning tidy tibbles ready for analysis.
 
 Developed by **Joash Joshua Ayo** (<joashjoshua789@gmail.com>).
 
-## Why brapiR2?
+## Supported BrAPI Modules
 
-| Feature | brapiR2 | QBMS |
-|----|----|----|
-| Design | Stateless, functional, pipeable | Stateful, menu-driven |
-| BrAPI v2 coverage | 32/36 entities, all 4 modules, read-only | See QBMS’s own documentation |
-| Genotyping support | Native variants, callsets, dosage matrix | Via GIGWA wrapper |
-| Return type | Always tibbles | Mixed lists/dataframes |
-| Auth | Unified token/OAuth2 | Engine-specific functions |
-| Caching | Built-in response caching | Limited |
+Every entity brapiR2 wraps, by module:
 
-**brapiR2 complements
-[QBMS](https://cran.r-project.org/package=QBMS)** - use QBMS for
-interactive exploration, use brapiR2 for programmatic pipelines and
-custom tooling.
+- **Core**: Programs, Trials, Studies, Locations, Seasons, Lists, People
+- **Germplasm**: Germplasm, Progeny, Attributes, Crosses, Seed Lots, and
+  Pedigree - both single-germplasm
+  ([`brapi_germplasm_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_germplasm_pedigree.md))
+  and batch/tree retrieval across many germplasm at once
+  ([`brapi_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_pedigree.md),
+  [`brapi_search_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_search_pedigree.md))
+- **Phenotyping**: Observation Units, Observations, Variables, Traits,
+  Scales, Methods, Ontologies, Images, Events
+- **Genotyping**: Samples, Variants, Variant Sets, Calls, Call Sets,
+  References, Allele Matrix, Genome Maps
 
-[brapir-v2](https://github.com/mverouden/brapir-v2) is another R client
-that, like brapiR2, targets the BrAPI v2 specification directly rather
-than v1. Its own README describes it as still under development, and the
-repository has had no commits in roughly four years (last pushed April
-2022) - it does not appear to be actively maintained.
-
-[BrAPI.R](https://github.com/TriticeaeToolbox/BrAPI.R) (David Waring,
-Cornell) is a different kind of tool entirely, and complementary rather
-than competing: its own DESCRIPTION calls it “simple wrapper functions
-for httr that make it easier to make manual HTTP calls to a BrAPI
-server”, and its README is explicit that it “does not have any knowledge
-of the currently supported BrAPI endpoints”. It’s a transport layer -
-callers pass endpoint paths as strings (`GET`, `POST`, and `PUT` are all
-supported, plus a two-step search helper) and get back raw nested lists,
-with a version argument that switches between BrAPI v1 and v2. It also
-ships Breedbase-specific functions explicitly outside the BrAPI spec.
-brapiR2 takes the opposite approach - named per-endpoint functions
-returning tibbles - and covers less ground on writes: BrAPI.R’s
-`POST`/`PUT` support covers exactly the write operations brapiR2
-deliberately omits.
+Not yet covered: Common Crop Names, Germplasm Attribute Values, Plates,
+and Vendor (lab/vendor order-tracking endpoints). brapiR2 is read-only
+by design - it does not implement BrAPI’s `POST`/`PUT` write endpoints.
 
 ## Installation
 
@@ -100,6 +83,27 @@ all_data <- brapi_fetch_parallel(
[…]
