### 2026-09-17 2ae6010
Fix two causes of brapi_study_data() failing on real trial data
Nested JSON objects are normalised by key order as they are parsed. Key order is not significant in JSON, but at least one server varies it between records in the same response, which made identical observations compare as distinct. brapi_study_data() also drops exact duplicates before pivoting.

The simplification step used unlist(), which dropped the zero-length elements pivot_wider() creates for unmeasured unit/trait combinations and returned a short column. Those cells are now NA and simplified columns are character throughout
[…]
