### 2026-09-17 00ba1fe
Send a user agent with every request
Requests were anonymous, so server operators had no way to see what was calling them. brapiR2 now sends a user agent carrying the package version, read from DESCRIPTION so it cannot drift, plus the httr2 and R versions; requests from continuous integration are marked. brapi_connection() gains a user_agent argument to override it, per the rOpenSci guidance that users should be able to.
