+++ DESCRIPTION
@@ -42,7 +42,8 @@ Imports:
     rlang (>= 1.1.0),
     tibble (>= 3.2.0),
     tidyr (>= 1.3.0),
-    tidyselect (>= 1.2.0)
+    tidyselect (>= 1.2.0),
+    utils
 Suggests:
     furrr (>= 0.3.0),
     future (>= 1.33.0),
