### 2026-09-17 07d3857
Send a user agent, and make login use the shared request builder
Requests were anonymous. brapiR2 now sends a user agent carrying the package version, read from DESCRIPTION so it cannot drift, plus the httr2 and R versions; requests from continuous integration are marked. brapi_connection() gains a user_agent argument to override it.

brapi_login() built its own request and hardcoded /brapi/, so authentication could not reach servers using another path and sent no user agent or Accept header. It now goes through brapi_req() like everything else. The OAuth token request identifies itself too.


[…]
