+++ NAMESPACE
@@ -87,6 +87,7 @@ importFrom(httr2,
   req_method,
   req_perform,
   req_retry,
+  req_timeout,
   req_url_path_append,
   req_url_query,
   request,
