+++ reference/brapi_cache_clear.md
@@ -25,9 +25,9 @@ Invisibly returns `con`.
 ``` r
 con <- brapi_connection("https://test-server.brapi.org")
 con <- brapi_cache_enable(con, dir = tempdir())
-#> ✔ Caching enabled at /tmp/RtmpnsZzi3 (TTL: 3600s)
+#> ✔ Caching enabled at /tmp/Rtmpix7SCn (TTL: 3600s)
 brapi_cache_clear(con)
-#> Warning: cannot remove file '/tmp/RtmpnsZzi3/bslib-e9b2b13fa612f50d23e4850d93d60d01', reason 'Directory not empty'
-#> Warning: cannot remove file '/tmp/RtmpnsZzi3/downlit', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/Rtmpix7SCn/bslib-e9b2b13fa612f50d23e4850d93d60d01', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/Rtmpix7SCn/downlit', reason 'Directory not empty'
 #> ✔ Cleared 7 cached response(s).
 ```
