### 2026-09-17 ef0d551
Trim whitespace around observation values
Values such as '80 ' would become NA on conversion to numeric. Reported by @dwaring87 from a development server; not reproducible across 21 studies on T3/Oat Sandbox, T3/Wheat Sandbox and Cassavabase, so the trim is defensive rather than a fix for observed data. ropensci/software-review#792
