+++ tests/testthat/test-phenotyping-mocked.R
@@ -137,3 +137,23 @@ test_that("brapi_study_data fills unmeasured traits with NA, not a short column"
   # u2 has no yield observation, so that cell must be NA rather than dropped.
   expect_true(is.na(out$yield[out$observationUnitDbId == "u2"]))
 })
+
+test_that("brapi_study_data trims whitespace around values", {
+  local_mocked_bindings(
+    brapi_observations = function(con, ...) {
+      tibble::tibble(
+        observationUnitDbId = c("u1", "u2"),
+        observationUnitName = c("p1", "p2"),
+        germplasmDbId = c("g1", "g1"),
+        germplasmName = c("G", "G"),
+        studyDbId = c("s1", "s1"),
+        observationVariableName = c("height", "height"),
+        value = c("150", " 80 ")
+      )
+    },
+    .package = "brapiR2"
+  )
+
+  out <- brapi_study_data(brapi_connection("https://example.org"), "s1")
+  expect_identical(sort(out$height), c("150", "80"))
+})
