### 2026-09-17 d268200
Add working notes on the servers surveyed so far
