### 2026-09-18 7212a07
Redirect the deprecated pedigree endpoints to /pedigree
BrAPI deprecated /germplasm/{id}/pedigree and /germplasm/{id}/progeny in v2.1. Both functions now query /pedigree?germplasmDbId= instead. This is not a transparent swap: the endpoints return different shapes, 15 columns rather than 8, with pedigree renamed to pedigreeString and relatives as tidy tibbles rather than raw nested lists. Verified on wheat-sandbox and oat-sandbox.

The public test server ignores the germplasmDbId filter on /pedigree, returning germplasm3 alongside germplasm1, so both functions now filter client-side as well.
