+++ NEWS.md
@@ -16,6 +16,15 @@
   `referenceName`/`start` are gone. Existing positional calls
   (`brapi_get_marker_map(con, variantSetDbId)`) still work, but code
   reading `referenceName` or `start` from the result will break.
+* `brapi_germplasm_pedigree()` and `brapi_germplasm_progeny()` now query
+  `/pedigree?germplasmDbId=` instead of the `/germplasm/{germplasmDbId}/pedigree`
+  and `/germplasm/{germplasmDbId}/progeny` sub-resources, which BrAPI
+  deprecated in v2.1. **The returned columns have changed**: 15 rather than 8,
+  `pedigree` is now `pedigreeString`, and `parents`, `siblings` and `progeny`
+  are list-columns of tidy tibbles rather than raw nested lists. The added
+  fields are `progeny`, `germplasmPUI`, `defaultDisplayName`,
+  `breedingMethodName`, `breedingMethodDbId`, `additionalInfo` and
+  `externalReferences` (@dwaring87, ropensci/software-review#792).
 
 ### Bug fixes
 
@@ -57,6 +66,10 @@
   server; not reproducible across 21 studies on T3/Oat Sandbox,
   T3/Wheat Sandbox and Cassavabase, so the trim is defensive
   (ropensci/software-review#792).
+* `brapi_germplasm_pedigree()` and `brapi_germplasm_progeny()` filter the
+  response on `germplasmDbId` client-side. The public test server ignores
+  that parameter on `/pedigree`, so a request for one germplasm returned
+  another alongside it (@dwaring87, ropensci/software-review#792).
 
 ### New features
 
