+++ articles/brapiR2.md
@@ -249,7 +249,7 @@ now reads from there instead:
 
 # Positions for every variant in the set, wherever they have been placed
 markers <- brapi_get_marker_map(con, variantSetDbId = vs_id)
-#> ℹ Async search started (ID: 5feed6d8-f7d3-4dc8-8567-40f17c1ad2ca). Polling...
+#> ℹ Async search started (ID: 4c2064a6-3b62-47fd-815d-8aa8b3374f0b). Polling...
 #> Warning: 14 of 20 variants in "variantset1" have no marker position record; returning
 #> positions for the remaining 6.
 markers
@@ -414,7 +414,7 @@ by default, per connection, and never shared between connections.
 cache_dir <- tempfile("brapi_cache_")
 dir.create(cache_dir)
 perf_con <- brapi_cache_enable(con, ttl = 3600, dir = cache_dir)
-#> ✔ Caching enabled at /tmp/Rtmp9oVTxl/brapi_cache_1e011dadf0f8 (TTL: 3600s)
+#> ✔ Caching enabled at /tmp/Rtmpa38EIV/brapi_cache_1f0d349af9c3 (TTL: 3600s)
 
 # First call: hits the server
 invisible(brapi_programs(perf_con))
