+++ reference/brapi_germplasm_pedigree.md
@@ -1,6 +1,8 @@
 # Get Germplasm Pedigree
 
-Get Germplasm Pedigree
+The `/germplasm/{germplasmDbId}/pedigree` endpoint this function
+originally called was deprecated in BrAPI v2.1. It now queries
+`/pedigree?germplasmDbId=` instead, which returns a richer record.
 
 ## Usage
 
@@ -22,7 +24,10 @@ brapi_germplasm_pedigree(con, germplasmDbId)
 
 ## Value
 
-A tibble with pedigree information (parents, crosses).
+A single-row tibble of the germplasm's pedigree node, with `parents`,
+`siblings` and `progeny` as list-columns of tidy tibbles. See
+[`brapi_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_pedigree.md),
+which this function calls.
 
 ## Examples
 
@@ -36,7 +41,7 @@ brapi_germplasm_pedigree(con, "germplasm1")
 #> 1 NA             NA                 breeding_method1   Male Backcross    
 #> # ℹ 11 more variables: crossingProjectDbId <chr>, crossingYear <int>,
 #> #   defaultDisplayName <chr>, familyCode <chr>, germplasmDbId <chr>,
-#> #   germplasmName <chr>, germplasmPUI <chr>, parents <lgl>,
-#> #   pedigreeString <chr>, progeny <lgl>, siblings <lgl>
+#> #   germplasmName <chr>, germplasmPUI <chr>, parents <list>,
+#> #   pedigreeString <chr>, progeny <list>, siblings <list>
 # }
 ```
