### 2026-09-18 1c7ed9b
Add brapi_location() for fetching a single location by ID
A user who knows a study's locationDbId had to list every location and filter to get its coordinates. Reported by @dwaring87 in ropensci/software-review#792. Verified against the public test server, Cassavabase, T3/Oat Sandbox and USDA-GRIN.
