+++ man/brapi_location.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/core.R
+\name{brapi_location}
+\alias{brapi_location}
+\title{Get a Single Location by ID}
+\usage{
+brapi_location(con, locationDbId)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{locationDbId}{Character. The unique location identifier.}
+}
+\value{
+A single-row tibble with location metadata, including
+coordinates where the server provides them.
+}
+\description{
+Get a Single Location by ID
+}
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+brapi_location(con, "location_01")
+}
+
+}
