+++ NAMESPACE
@@ -21,6 +21,7 @@ export(brapi_get_dosage_matrix)
 export(brapi_get_marker_map)
 export(brapi_images)
 export(brapi_lists)
+export(brapi_location)
 export(brapi_locations)
 export(brapi_login)
 export(brapi_login_oauth2)
