### 2026-09-18 b609485
Add brapi_list(), and stop mistaking a data field for a record envelope
brapi_lists() returns only metadata, so there was no way to reach a list's contents. brapi_list() returns a single list with its members as a character vector in the data list-column. Reported by @dwaring87 in ropensci/software-review#792.

This needed a parser fix first: any result\ was treated as the record envelope, so /lists/{listDbId} returned three member strings and discarded the other twelve fields. data is now the envelope only when empty or holding objects, verified against thirteen endpoint responses on the test
[…]
