+++ man/brapi_list.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/core.R
+\name{brapi_list}
+\alias{brapi_list}
+\title{Get a Single List by ID, With Its Contents}
+\usage{
+brapi_list(con, listDbId)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{listDbId}{Character. The unique list identifier.}
+}
+\value{
+A single-row tibble of list metadata, with the list's members
+as a character vector in the \code{data} list-column.
+}
+\description{
+Unlike \code{\link[=brapi_lists]{brapi_lists()}}, which returns only list metadata, this returns
+a single list together with its members in the \code{data} list-column.
+\code{listType} says what the members are (for example \code{"germplasm"}).
+}
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+lst <- brapi_list(con, "list1")
+lst$data[[1]]
+}
+
+}
