+++ man/brapi_shared_ids.Rd
@@ -17,6 +17,8 @@
 \item{ontologyDbId}{Character. The unique ontology identifier.}
 
 \item{locationDbId}{Character. The unique location identifier.}
+
+\item{listDbId}{Character. The unique list identifier.}
 }
 \description{
 Not a real function - like \link{brapi_shared_params}, this exists only as an
