+++ NAMESPACE
@@ -20,6 +20,7 @@ export(brapi_germplasm_progeny)
 export(brapi_get_dosage_matrix)
 export(brapi_get_marker_map)
 export(brapi_images)
+export(brapi_list)
 export(brapi_lists)
 export(brapi_location)
 export(brapi_locations)
