### 2026-09-18 aaacd72
Add brapi_location() and brapi_list() to the pkgdown reference index
