+++ _pkgdown.yml
@@ -25,8 +25,10 @@ reference:
       - brapi_studies
       - brapi_study
       - brapi_locations
+      - brapi_location
       - brapi_seasons
       - brapi_lists
+      - brapi_list
       - brapi_people
       - brapi_server_info
 
