### 2026-09-19 52ef469
Report the server's own error message, not just the HTTP status
BrAPI servers explain failures in several different ways, and brapiR2 discarded all of them in favour of the status code. Requests no longer let httr2 raise the error; the response body is read first and whatever message it carries is included.

Handles the specification's metadata.status envelope (Breedbase and T3, including on HTTP 200 where a bad password returns a null token), a bare JSON Message field (GRIN-Global), and a plain-text sentence (the public test server). HTML error pages and empty bodies fall back to the status a
[…]
