+++ DESCRIPTION
@@ -40,6 +40,7 @@ Imports:
     jsonlite (>= 1.8.0),
     purrr (>= 1.0.0),
     rlang (>= 1.1.0),
+    stats,
     tibble (>= 3.2.0),
     tidyr (>= 1.3.0),
     tidyselect (>= 1.2.0),
