### 2026-09-19 403dce8
Export brapi_get() and brapi_post_search()
The named functions cover 32 of the 36 BrAPI v2.1 entities. These are the layer beneath them, for endpoints brapiR2 does not wrap, server extensions, and query parameters no named function exposes. Pagination, caching, authentication and error reporting work as they do for the named functions. Recommended by both reviewers and by the rOpenSci packaging guidelines.

Exporting brapi_get() immediately exposed a parser bug: /commoncropnames returns its crop names as bare strings inside result\, and the whole response came back as a single row with a list-
[…]
