+++ man/brapi_post_search.Rd
@@ -1,0 +1,64 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/request.R
+\name{brapi_post_search}
+\alias{brapi_post_search}
+\title{Call Any BrAPI Search Endpoint}
+\usage{
+brapi_post_search(
+  con,
+  endpoint,
+  body = list(),
+  poll_interval = 2,
+  max_polls = 30L
+)
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{endpoint}{Character. The search endpoint, with or without a
+leading slash (for example \code{"/search/germplasm"}).}
+
+\item{body}{Named list. The search request body. Filter fields are sent
+as JSON arrays, as BrAPI expects, even when you supply a single value.}
+
+\item{poll_interval}{Numeric. Seconds between polling attempts for an
+asynchronous search. Default 2.}
+
+\item{max_polls}{Integer. Maximum polling attempts before giving up.
+Default 30.}
+}
+\value{
+A tibble of search results.
+}
+\description{
+The companion to \code{\link[=brapi_get]{brapi_get()}} for the POST \verb{/search/\{entity\}}
+endpoints, which take a filter body rather than query parameters and
+may run asynchronously. Use it for search endpoints brapiR2 does not
+wrap, or for filter fields a named search function does not expose.
+}
+\section{Asynchronous searches}{
+
+A server may answer immediately with the results, or with HTTP 202 and
+a \code{searchResultsDbId} to be polled until the results are ready. Both
+are handled here; the polling happens inside the call and you get the
+finished results either way.
+}
+
+\section{Return shape}{
+
+As with \code{\link[=brapi_get]{brapi_get()}}, the response passes through the same parser the
+named functions use. A well-formed BrAPI result returns one row per
+record; an unusual one may need reshaping yourself.
+}
+
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+brapi_post_search(con, "/search/germplasm",
+                  body = list(germplasmNames = "Tomatillo Fantastico"))
+}
+
+}
+\seealso{
+\code{\link[=brapi_get]{brapi_get()}} for the GET endpoints.
+}
