+++ man/brapi_get.Rd
@@ -1,0 +1,54 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/request.R
+\name{brapi_get}
+\alias{brapi_get}
+\title{Call Any BrAPI GET Endpoint}
+\usage{
+brapi_get(con, endpoint, query = list())
+}
+\arguments{
+\item{con}{A \code{\link[=brapi_connection]{brapi_connection()}} object.}
+
+\item{endpoint}{Character. The endpoint path, with or without a leading
+slash (for example \code{"/programs"} or \code{"commoncropnames"}). The base
+URL, BrAPI path and version come from \code{con}.}
+
+\item{query}{Named list. Query parameters to append to the URL.
+\code{pageSize} defaults to the connection's page size; \code{page} is managed
+by the pagination loop and should not be set here.}
+}
+\value{
+A tibble of results, or an empty tibble if the endpoint
+returned no data.
+}
+\description{
+The named functions in brapiR2 cover 32 of the 36 BrAPI v2.1 entities.
+This is the layer beneath them, for endpoints brapiR2 does not wrap, for
+servers with non-standard extensions, and for query parameters a named
+function does not expose. Pagination, caching, authentication and error
+reporting work exactly as they do for the named functions.
+}
+\section{Return shape}{
+
+The response passes through the same parser the named functions use, so
+a well-formed BrAPI collection returns one row per record. An endpoint
+returning something the parser does not recognise may come back with
+list-columns or a shape you need to reshape yourself. The named
+functions are the better choice wherever one exists.
+}
+
+\examples{
+\donttest{
+con <- brapi_connection("https://test-server.brapi.org")
+
+# An endpoint brapiR2 does not wrap
+brapi_get(con, "/commoncropnames")
+
+# A query parameter no named function exposes
+brapi_get(con, "/studies", query = list(active = "true"))
+}
+
+}
+\seealso{
+\code{\link[=brapi_post_search]{brapi_post_search()}} for the POST search endpoints.
+}
