+++ news/index.md
@@ -101,6 +101,9 @@
   only its members and discarded every other field. `data` is now
   treated as the envelope only when it is absent of scalars — that is,
   empty or holding objects.
+- A collection whose records are bare strings is no longer mistaken for
+  a single object. `/commoncropnames` returns its crop names that way,
+  and the whole response came back as one row with a list-column.
 
 #### Minor improvements
 
@@ -182,6 +185,16 @@
   the `data` list-column, ready to pass to another function
   ([@dwaring87](https://github.com/dwaring87),
   ropensci/software-review#792).
+- [`brapi_get()`](https://josh45-source.github.io/brapiR2/reference/brapi_get.md)
+  and
+  [`brapi_post_search()`](https://josh45-source.github.io/brapiR2/reference/brapi_post_search.md)
+  are now exported, so an endpoint brapiR2 does not wrap, a server
+  extension, or a query parameter no named function exposes can be
+  reached without dropping to raw HTTP. Pagination, caching,
+  authentication and error reporting work as they do for the named
+  functions. Recommended by both reviewers and by the rOpenSci packaging
+  guidelines ([@dwaring87](https://github.com/dwaring87),
+  [@jmh579](https://github.com/jmh579), ropensci/software-review#792).
 
 #### Deprecated
 
