### 2026-09-19 205f113
Verify that observations really are from the requested study
A server may accept the studyDbId filter and ignore it; the public test server does exactly this on /pedigree. brapi_study_data() would then have pivoted every study's observations together. Extra studies are filtered out client-side and the function reports how many it found. Suggested by @dwaring87 in ropensci/software-review#792.
