### 2026-09-19 80936de
Record the README install changes and this round's tests in NEWS
The pak::pak() switch and the build_vignettes note were made but never recorded. The pedigree entry still described brapi_germplasm_pedigree() as a parallel function rather than one that now delegates. The Testing section described only the pre-review work.
