+++ news/index.md
@@ -139,14 +139,18 @@
   [`brapi_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_pedigree.md)
   and
   [`brapi_search_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_search_pedigree.md)
-  retrieve pedigree records across many germplasm in one call (via
-  `/pedigree` and `/search/pedigree`), complementing the existing
-  single-germplasm
-  [`brapi_germplasm_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_germplasm_pedigree.md).
-  Each row is one pedigree node; `parents`, `siblings`, and `progeny`,
-  when requested, are list-columns of tidy per-node tibbles (not raw
-  nested lists), so a marker on several maps or a node with several
-  relatives is never silently collapsed to one row.
+  retrieve pedigree records across many germplasm in one call, via
+  `/pedigree` and `/search/pedigree`.
+  [`brapi_germplasm_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_germplasm_pedigree.md)
+  and
+  [`brapi_germplasm_progeny()`](https://josh45-source.github.io/brapiR2/reference/brapi_germplasm_progeny.md)
+  now delegate to
+  [`brapi_pedigree()`](https://josh45-source.github.io/brapiR2/reference/brapi_pedigree.md)
+  rather than calling the sub-resources BrAPI deprecated in v2.1. Each
+  row is one pedigree node; `parents`, `siblings` and `progeny`, when
+  requested, are list-columns of tidy per-node tibbles rather than raw
+  nested lists, so a node with several relatives is never silently
+  collapsed to one row.
 - New Ontologies entity support (`R/phenotyping.R`):
   [`brapi_ontologies()`](https://josh45-source.github.io/brapiR2/reference/brapi_ontologies.md)
   and
@@ -253,6 +257,14 @@
   corrected, and `inst/WORDLIST` has been extended with the domain
   vocabulary and package names the spellchecker cannot know
   ([@jmh579](https://github.com/jmh579), ropensci/software-review#792).
+- The README installs with
+  [`pak::pak()`](https://pak.r-lib.org/reference/pak.html) rather than
+  `remotes::install_github()`, which now warns
+  ([@jmh579](https://github.com/jmh579), ropensci/software-review#792),
+  and notes that `remotes` and `devtools` need `build_vignettes = TRUE`
+  for the vignette to be installed at all
+  ([@dwaring87](https://github.com/dwaring87),
+  ropensci/software-review#792).
 - The README leads with what brapiR2 does and which BrAPI modules it
   covers. The QBMS comparison table and the notes on other BrAPI clients
   have moved into Related Packages, and Authentication now comes before
@@ -261,12 +273,20 @@
 
 #### Testing
 
-- Substantially expanded the mocked and live-server integration test
-  suites alongside the features above: argument-capturing tests for
-  every new thin wrapper, dedicated tests for the pedigree relative-list
-  parsing (nodes with parents, with progeny, and with neither), and
-  guarded integration tests against the public BrAPI test server for
-  every new function.
+- Substantially expanded the mocked and live-server test suites
+  alongside the features above: argument-capturing tests for every new
+  thin wrapper, dedicated tests for the pedigree relative-list parsing
+  (nodes with parents, with progeny, and with neither), and guarded
+  integration tests against the public BrAPI test server for every new
+  function.
+- Added tests for the work done in response to review: URL construction
+  and cache keys under a non-default `path`, the user agent as sent and
+  as overridden, login reaching a non-standard path, JSON key-order
[…]
