### 2026-09-19 c6677d1
Show rather than run the alternative GS model fits
The genomic selection article fitted the same trait three ways; the BGLR and sommer results were printed and never used again. The comparative prose on when to reach for each package stays, but the chunk is no longer evaluated, which removes two heavy dependencies from the build. Raised by @jmh579 in ropensci/software-review#792.
