### 2026-09-19 2572de7
Document each function's BrAPI endpoint and its query parameters
Every wrapper's help page now names the endpoint it calls, links to that endpoint's definition in the v2.1 specification, and lists the query parameters the specification defines for it. brapi_studies() documented only trialDbId while the specification defines nine; that gap is now visible on every function. Asked for by @dwaring87 and @jmh579 in ropensci/software-review#792.

All of it is derived rather than transcribed. dev/brapi-spec.R reads the specification from the pinned V2.1 tag of plantbreeding/BrAPI, dev/brapi-doc-block
[…]
