+++ R/germplasm.R
@@ -11,6 +11,16 @@
 #'
 #' @return A tibble with one row per germplasm accession.
 #'
+#' @section BrAPI endpoint:
+#' `GET /germplasm` - see the
+#' [v2.1 specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Germplasm/Germplasm/Germplasm_GET_POST.yaml).
+#'
+#' Query parameters the specification defines, which may be passed
+#' through `...`:
+#'
+#' `accessionNumber`, `collection`, `binomialName`, `genus`, `species`,
+#' `synonym`, `parentDbId`, `progenyDbId`.
+#'
 #' @examples
 #' \donttest{
 #' con <- brapi_connection("https://test-server.brapi.org")
@@ -29,6 +39,10 @@ brapi_germplasm <- function(con, ...) {
 #' @inheritParams brapi_shared_ids
 #'
 #' @return A single-row tibble with germplasm details.
+#'
+#' @section BrAPI endpoint:
+#' `GET /germplasm/{germplasmDbId}` - see the
+#' [v2.1 specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Germplasm/Germplasm/Germplasm_GermplasmDbId_GET_PUT.yaml).
 #'
 #' @examples
 #' \donttest{
@@ -54,6 +68,17 @@ brapi_germplasm_detail <- function(con, germplasmDbId) {
 #' @return A single-row tibble of the germplasm's pedigree node, with
 #'   `parents`, `siblings` and `progeny` as list-columns of tidy tibbles.
 #'   See [brapi_pedigree()], which this function calls.
+#'
+#' @section BrAPI endpoint:
+#' `GET /pedigree` - see the
+#' [v2.1 specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Germplasm/Pedigree/Pedigree_GET_POST_PUT.yaml).
+#'
+#' Query parameters the specification defines, which may be passed
+#' through `...`:
+#'
+#' `accessionNumber`, `collection`, `familyCode`, `binomialName`, `genus`,
+#' `species`, `synonym`, `includeParents`, `includeSiblings`,
+#' `includeProgeny`, `includeFullTree`, `pedigreeDepth`, `progenyDepth`.
 #'
 #' @examples
 #' \donttest{
@@ -83,6 +108,17 @@ brapi_germplasm_pedigree <- function(con, germplasmDbId) {
 #'   `progeny` as a list-column of a tidy tibble of descendants. See
 #'   [brapi_pedigree()], which this function calls.
 #'
+#' @section BrAPI endpoint:
+#' `GET /pedigree` - see the
+#' [v2.1 specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Germplasm/Pedigree/Pedigree_GET_POST_PUT.yaml).
+#'
+#' Query parameters the specification defines, which may be passed
+#' through `...`:
+#'
+#' `accessionNumber`, `collection`, `familyCode`, `binomialName`, `genus`,
+#' `species`, `synonym`, `includeParents`, `includeSiblings`,
+#' `includeProgeny`, `includeFullTree`, `pedigreeDepth`, `progenyDepth`.
+#'
 #' @examples
 #' \donttest{
 #' con <- brapi_connection("https://test-server.brapi.org")
@@ -109,6 +145,15 @@ brapi_germplasm_progeny <- function(con, germplasmDbId) {
 #'
 #' @return A tibble with one row per attribute definition.
 #'
+#' @section BrAPI endpoint:
+#' `GET /attributes` - see the
[…]
