+++ R/utils.R
@@ -9,6 +9,15 @@
 #'
 #' @return Logical. `TRUE` if the server responds, `FALSE` otherwise.
 #'
+#' @section BrAPI endpoint:
+#' `GET /serverinfo` - see the
+#' [v2.1 specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Core/ServerInfo/ServerInfo_GET.yaml).
+#'
+#' Query parameters the specification defines, which may be passed
+#' through `...`:
+#'
+#' `contentType`, `dataType`.
+#'
 #' @examples
 #' \donttest{
 #' con <- brapi_connection("https://test-server.brapi.org")
@@ -47,6 +56,15 @@ brapi_ping <- function(con) {
 #'
 #' @return A tibble with columns for endpoint service, method(s), and
 #'   version(s).
+#'
+#' @section BrAPI endpoint:
+#' `GET /serverinfo` - see the
+#' [v2.1 specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Core/ServerInfo/ServerInfo_GET.yaml).
+#'
+#' Query parameters the specification defines, which may be passed
+#' through `...`:
+#'
+#' `contentType`, `dataType`.
 #'
 #' @examples
 #' \donttest{
