+++ reference/brapi_reference_sets.md
@@ -24,6 +24,16 @@ brapi_reference_sets(con, ...)
 
 A tibble with one row per reference set.
 
+## BrAPI endpoint
+
+`GET /referencesets` - see the [v2.1
+specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/BrAPI-Genotyping/ReferenceSets/ReferenceSets_GET.yaml).
+
+Query parameters the specification defines, which may be passed through
+`...`:
+
+`referenceSetDbId`, `accession`, `assemblyPUI`, `md5checksum`.
+
 ## Examples
 
 ``` r
