### 2026-09-19 7092549
Record the endpoint documentation and the corrected coverage figures in NEWS
