+++ reference/brapi_search_calls.md
@@ -43,7 +43,7 @@ specification](https://github.com/plantbreeding/BrAPI/blob/V2.1/Specification/Br
 # \donttest{
 con <- brapi_connection("https://test-server.brapi.org")
 brapi_search_calls(con, variantSetDbIds = "variantset1")
-#> ℹ Async search started (ID: 908a524f-c9a0-457c-a050-627544c24ee9). Polling...
+#> ℹ Async search started (ID: 5fbe59a8-e8fd-4ecd-be9d-14c766bbaf25). Polling...
 #> # A tibble: 260 × 12
 #>    additionalInfo callSetDbId callSetName genotype         genotypeValue
 #>    <lgl>          <chr>       <chr>       <list>           <chr>        
