### 2026-09-19 8b2e090
Add ?brapi_coverage listing every wrapped BrAPI endpoint
brapiR2 wraps 49 of the specification's 138 retrieval endpoints, and until now there was nowhere to see which. The topic groups them by module with the function that wraps each one, names the five entities with no wrapper, and distinguishes itself from brapi_endpoints(), which asks a server what it implements. Generated by dev/brapi-coverage.R from the specification inventory. Asked for by @jmh579 and @dwaring87 in ropensci/software-review#792.
