+++ dev/brapi-coverage.R
@@ -1,0 +1,64 @@
+# Generate R/coverage.R, the ?brapi_coverage help topic.
+#
+# Reads dev/brapi-endpoints.rds (from dev/brapi-spec.R) and
+# dev/brapi-function-map.rds, and writes a documentation-only topic
+# listing every wrapped endpoint by module with the function that wraps
+# it. Re-run after adding or removing a wrapper.
+
+e <- readRDS("dev/brapi-endpoints.rds")
+m <- readRDS("dev/brapi-function-map.rds")
+
+m$method <- ifelse(grepl("^/search/", m$path), "POST", "GET")
+k <- paste(e$path, e$method)
+m$module <- e$module[match(paste(m$path, m$method), k)]
+m <- m[order(m$module, m$path, m$fn), ]
+
+mods <- c("Core", "Germplasm", "Phenotyping", "Genotyping")
+
+lines <- c(
+  "# Generated by dev/brapi-coverage.R - do not edit by hand.",
+  "",
+  "#' BrAPI Endpoints Covered by brapiR2",
+  "#'",
+  paste0(
+    "#' brapiR2 wraps ", length(unique(m$path)),
+    " of the 138 retrieval endpoints the BrAPI v2.1"
+  ),
+  "#' specification defines, across 32 of its 37 top-level entities. This",
+  "#' topic lists them by module, with the function that wraps each one.",
+  "#'",
+  "#' Each function's own help page names its endpoint, links to the",
+  "#' specification, and lists the query parameters that endpoint accepts.",
+  "#' [brapi_endpoints()] is a different thing: it asks a particular server",
+  "#' which endpoints *it* implements, which is usually a smaller set again.",
+  "#'"
+)
+
+for (mod in mods) {
+  sub <- m[m$module == mod, ]
+  lines <- c(lines, paste0("#' @section ", mod, " module:"), "#' \\describe{")
+  for (p in unique(sub$path)) {
+    fns <- sub$fn[sub$path == p]
+    meth <- sub$method[sub$path == p][1]
+    lines <- c(lines, paste0(
+      "#'   \\item{`", meth, " ", p, "`}{",
+      paste0("[", fns, "()]", collapse = ", "), "}"
+    ))
+  }
+  lines <- c(lines, "#' }", "#'")
+}
+
+lines <- c(
+  lines,
+  "#' @section Not covered:",
+  "#' Common Crop Names, Germplasm Attribute Values, Planned Crosses,",
+  "#' Plates and Vendor Samples have no wrapper. `POST` and `PUT` write",
+  "#' endpoints are out of scope by design. [brapi_get()] and",
+  "#' [brapi_post_search()] reach anything not wrapped here.",
+  "#'",
+  "#' @name brapi_coverage",
+  "NULL"
+)
+
+writeLines(lines, "R/coverage.R")
+message("Wrote R/coverage.R: ", length(unique(m$path)), " endpoints.")
