+++ R/coverage.R
@@ -1,0 +1,86 @@
+# Generated by dev/brapi-coverage.R - do not edit by hand.
+
+#' BrAPI Endpoints Covered by brapiR2
+#'
+#' brapiR2 wraps 49 of the 138 retrieval endpoints the BrAPI v2.1
+#' specification defines, across 32 of its 37 top-level entities. This
+#' topic lists them by module, with the function that wraps each one.
+#'
+#' Each function's own help page names its endpoint, links to the
+#' specification, and lists the query parameters that endpoint accepts.
+#' [brapi_endpoints()] is a different thing: it asks a particular server
+#' which endpoints *it* implements, which is usually a smaller set again.
+#'
+#' @section Core module:
+#' \describe{
+#'   \item{`GET /lists`}{[brapi_lists()]}
+#'   \item{`GET /lists/{listDbId}`}{[brapi_list()]}
+#'   \item{`GET /locations`}{[brapi_locations()]}
+#'   \item{`GET /locations/{locationDbId}`}{[brapi_location()]}
+#'   \item{`GET /people`}{[brapi_people()]}
+#'   \item{`GET /programs`}{[brapi_programs()]}
+#'   \item{`GET /programs/{programDbId}`}{[brapi_program()]}
+#'   \item{`GET /seasons`}{[brapi_seasons()]}
+#'   \item{`GET /serverinfo`}{[brapi_endpoints()], [brapi_ping()], [brapi_server_info()]}
+#'   \item{`GET /studies`}{[brapi_studies()]}
+#'   \item{`GET /studies/{studyDbId}`}{[brapi_study()]}
+#'   \item{`GET /trials`}{[brapi_trials()]}
+#'   \item{`GET /trials/{trialDbId}`}{[brapi_trial()]}
+#' }
+#'
+#' @section Germplasm module:
+#' \describe{
+#'   \item{`GET /attributes`}{[brapi_germplasm_attributes()]}
+#'   \item{`GET /crosses`}{[brapi_crosses()]}
+#'   \item{`GET /crossingprojects`}{[brapi_crossing_projects()]}
+#'   \item{`GET /germplasm`}{[brapi_germplasm()]}
+#'   \item{`GET /germplasm/{germplasmDbId}`}{[brapi_germplasm_detail()]}
+#'   \item{`GET /pedigree`}{[brapi_germplasm_pedigree()], [brapi_germplasm_progeny()], [brapi_pedigree()]}
+#'   \item{`POST /search/germplasm`}{[brapi_search_germplasm()]}
+#'   \item{`POST /search/pedigree`}{[brapi_search_pedigree()]}
+#'   \item{`GET /seedlots`}{[brapi_seed_lots()]}
+#' }
+#'
+#' @section Phenotyping module:
+#' \describe{
+#'   \item{`GET /events`}{[brapi_events()]}
+#'   \item{`GET /images`}{[brapi_images()]}
+#'   \item{`GET /methods`}{[brapi_methods()]}
+#'   \item{`GET /observations`}{[brapi_observations()]}
+#'   \item{`GET /observationunits`}{[brapi_observation_units()]}
+#'   \item{`GET /ontologies`}{[brapi_ontologies()]}
+#'   \item{`GET /ontologies/{ontologyDbId}`}{[brapi_ontology()]}
+#'   \item{`GET /scales`}{[brapi_scales()]}
+#'   \item{`POST /search/observations`}{[brapi_search_observations()]}
+#'   \item{`POST /search/variables`}{[brapi_search_variables()]}
+#'   \item{`GET /traits`}{[brapi_traits()]}
+#'   \item{`GET /variables`}{[brapi_observation_variables()]}
[…]
