+++ reference/brapi_coverage.md
@@ -1,0 +1,231 @@
+# BrAPI Endpoints Covered by brapiR2
+
+brapiR2 wraps 49 of the 138 retrieval endpoints the BrAPI v2.1
+specification defines, across 32 of its 37 top-level entities. This
+topic lists them by module, with the function that wraps each one.
+
+## Details
+
+Each function's own help page names its endpoint, links to the
+specification, and lists the query parameters that endpoint accepts.
+[`brapi_endpoints()`](https://josh45-source.github.io/brapiR2/reference/brapi_endpoints.md)
+is a different thing: it asks a particular server which endpoints *it*
+implements, which is usually a smaller set again.
+
+## Core module
+
+- `GET /lists`:
+
+  [`brapi_lists()`](https://josh45-source.github.io/brapiR2/reference/brapi_lists.md)
+
+- `GET /lists/{listDbId}`:
+
+  [`brapi_list()`](https://josh45-source.github.io/brapiR2/reference/brapi_list.md)
+
+- `GET /locations`:
+
+  [`brapi_locations()`](https://josh45-source.github.io/brapiR2/reference/brapi_locations.md)
+
+- `GET /locations/{locationDbId}`:
+
+  [`brapi_location()`](https://josh45-source.github.io/brapiR2/reference/brapi_location.md)
+
+- `GET /people`:
+
+  [`brapi_people()`](https://josh45-source.github.io/brapiR2/reference/brapi_people.md)
+
+- `GET /programs`:
+
+  [`brapi_programs()`](https://josh45-source.github.io/brapiR2/reference/brapi_programs.md)
+
+- `GET /programs/{programDbId}`:
+
+  [`brapi_program()`](https://josh45-source.github.io/brapiR2/reference/brapi_program.md)
+
+- `GET /seasons`:
+
+  [`brapi_seasons()`](https://josh45-source.github.io/brapiR2/reference/brapi_seasons.md)
+
+- `GET /serverinfo`:
+
+  [`brapi_endpoints()`](https://josh45-source.github.io/brapiR2/reference/brapi_endpoints.md),
+  [`brapi_ping()`](https://josh45-source.github.io/brapiR2/reference/brapi_ping.md),
+  [`brapi_server_info()`](https://josh45-source.github.io/brapiR2/reference/brapi_server_info.md)
+
+- `GET /studies`:
+
+  [`brapi_studies()`](https://josh45-source.github.io/brapiR2/reference/brapi_studies.md)
+
+- `GET /studies/{studyDbId}`:
+
+  [`brapi_study()`](https://josh45-source.github.io/brapiR2/reference/brapi_study.md)
+
+- `GET /trials`:
+
+  [`brapi_trials()`](https://josh45-source.github.io/brapiR2/reference/brapi_trials.md)
+
+- `GET /trials/{trialDbId}`:
+
+  [`brapi_trial()`](https://josh45-source.github.io/brapiR2/reference/brapi_trial.md)
+
+## Germplasm module
+
+- `GET /attributes`:
+
+  [`brapi_germplasm_attributes()`](https://josh45-source.github.io/brapiR2/reference/brapi_germplasm_attributes.md)
+
+- `GET /crosses`:
+
+  [`brapi_crosses()`](https://josh45-source.github.io/brapiR2/reference/brapi_crosses.md)
+
+- `GET /crossingprojects`:
+
+  [`brapi_crossing_projects()`](https://josh45-source.github.io/brapiR2/reference/brapi_crossing_projects.md)
+
+- `GET /germplasm`:
+
+  [`brapi_germplasm()`](https://josh45-source.github.io/brapiR2/reference/brapi_germplasm.md)
+
[…]
