+++ news/index.md
@@ -292,6 +292,13 @@
   it: the specification defines 37 top-level entities rather than 36,
   brapiR2 wraps 49 retrieval endpoints rather than 56, and Planned
   Crosses was missing from the list of uncovered entities.
+- New
+  [`?brapi_coverage`](https://josh45-source.github.io/brapiR2/reference/brapi_coverage.md)
+  lists every BrAPI endpoint brapiR2 wraps, grouped by module, with the
+  function that wraps each one, and names the entities that have no
+  wrapper. Asked for by [@jmh579](https://github.com/jmh579) and
+  [@dwaring87](https://github.com/dwaring87)
+  (ropensci/software-review#792).
 
 #### Testing
 

[…]
