### 2026-09-19 c79bce7
Mark brapi_coverage as an internal topic for pkgcheck
It is a documentation-only topic with no function behind it, but pkgcheck reads it as an undocumented, exampleless function. @keywords internal is how a non-function topic is excluded; the README links it directly and it stays in the pkgdown reference index.
