+++ R/coverage.R
@@ -83,4 +83,5 @@
 #' [brapi_post_search()] reach anything not wrapped here.
 #'
 #' @name brapi_coverage
+#' @keywords internal
 NULL
