### 2026-09-19 5c579ad
Survey brapiR2 against every reachable BrAPI server
dev/survey-servers.R runs one request against each of ten endpoints on eight servers across three implementations and records the result. 62 of 80 probes succeeded; the failures are timeouts on large Breedbase instances, HTTP 500s on two of them, and endpoints GRIN-Global correctly does not implement.
