+++ dev/SERVERS.md
@@ -1,24 +1,54 @@
-# Server survey notes
+# Server survey
 
-Working notes for the compatibility survey (Tier 5). Not part of the
-package; `dev/` is .Rbuildignored.
+Working notes for the rOpenSci review. Not part of the package; `dev/`
+is `.Rbuildignore`d.
+
+`dev/survey-servers.R` runs one request against each of ten endpoints on
+every reachable server and records what happened, to `dev/server-survey.rds`.
+One request per endpoint, not a paginated fetch: the question is whether a
+server answers, not how much data it holds.
+
+Last run: 19 September 2026. 62 of 80 probes succeeded.
 
 ## Servers
 
-| Server | Base URL | Path | Auth for reads | Services | Studies |
-|---|---|---|---|---|---|
-| BrAPI test server | test-server.brapi.org | brapi | no | 148 | 3 |
-| Cassavabase | cassavabase.org | brapi | no | 118 | 8537 |
-| T3/Oat Sandbox | oat-sandbox.triticeaetoolbox.org | brapi | yes | 118 | 2624 |
-| T3/Wheat Sandbox | wheat-sandbox.triticeaetoolbox.org | brapi | yes | 118 | 9030 |
-| USDA-GRIN | npgsweb.ars-grin.gov | gringlobal/brapi | no | 13 | n/a |
+| Server | Implementation | Auth for reads | Services | Probes passed |
+|---|---|---|---|---|
+| BrAPI test server | reference | no | 148 | 10/10 |
+| Cassavabase | Breedbase | no | 118 | 7/10 |
+| Sweetpotatobase | Breedbase | no | 118 | 9/10 |
+| Coffeabase | Breedbase | no | 117 | 10/10 |
+| Citrusgreening | Breedbase | no | 117 | 8/10 |
+| USDA-GRIN | GRIN-Global | no | 13 | 2/10 |
+| T3/Oat Sandbox | Breedbase | yes | 118 | 8/10 |
+| T3/Wheat Sandbox | Breedbase | yes | 118 | 8/10 |
 
-T3 accounts are per-instance; oat credentials do not work on wheat.
-T3 requires authentication even for /serverinfo, which the spec treats
-as the discovery endpoint. Cassavabase and GRIN do not.
+T3 accounts are per instance; oat credentials do not work on wheat.
+T3 requires authentication even for `/serverinfo`, which the specification
+treats as the discovery endpoint. The others do not.
 
-Login worked first time on both T3 sandboxes, so the Breedbase login
-bug @dwaring87 and @jmh579 hit on Cassavabase is not present there.
+## What failed
+
+| Server | Endpoint | Seconds | Result |
+|---|---|---|---|
+| Cassavabase | `/samples` | 30.0 | no response within the 30s timeout |
+| Cassavabase | `/trials` | 30.0 | no response within the 30s timeout |
+| Cassavabase | `/variantsets` | 30.0 | no response within the 30s timeout |
+| Citrusgreening | `/studies` | 1.3 | BrAPI request failed (HTTP 500). |
+| Citrusgreening | `/variantsets` | 1.3 | BrAPI request failed (HTTP 500). |
+| Sweetpotatobase | `/samples` | 1.6 | BrAPI request failed (HTTP 500). |
+| T3/Oat Sandbox | `/samples` | 21.1 | could not connect (transient) |
+| T3/Oat Sandbox | `/variantsets` | 30.0 | no response within the 30s timeout |
+| T3/Wheat Sandbox | `/trials` | 30.0 | no response within the 30s timeout |
+| T3/Wheat Sandbox | `/variantsets` | 30.1 | no response within the 30s timeout |
[…]
