### 2026-09-19 64eebe9
Claim the specification rather than five untested databases
DESCRIPTION named Breedbase, BMS, EBS, GIGWA and Germinate; only Breedbase had ever been tested. It now claims BrAPI v2 compliance, and the README records the eight servers across three implementations that brapiR2 has been exercised against, the 62 of 80 endpoint probes that succeeded, and the four databases that remain untested for want of access. Raised by @jmh579 in ropensci/software-review#792.
