+++ DESCRIPTION
@@ -21,8 +21,8 @@ Description: Provides pipe-friendly, stateless read access to the Breeding
     search handling, response caching, parallel batch fetching, and
     convenience functions for genomic selection workflows (e.g. dosage
     matrix extraction). Designed for plant breeders and bioinformaticians
-    who need programmatic access to 'Breedbase', 'BMS', 'EBS', 'GIGWA',
-    'Germinate', and any 'BrAPI'-compliant server.
+    who need programmatic access to plant breeding databases that
+    implement the 'BrAPI' v2 specification.
 License: MIT + file LICENSE
 URL: https://github.com/josh45-source/brapiR2,
     https://josh45-source.github.io/brapiR2/
