### 2026-09-19 93c1df5
Add max_pages to brapi_get() for looking at large servers
An unfiltered call walks every page. Cassavabase took 18 minutes to return 8,539 studies and a germplasm listing on T3/Wheat ran over an hour, with no way to ask for just the first page. The default is unchanged and a truncated result is never cached, so a later full call cannot read a partial one back as complete. Found while surveying public servers.
