+++ NEWS.md
@@ -145,6 +145,12 @@
   caching, authentication and error reporting work as they do for the named
   functions. Recommended by both reviewers and by the rOpenSci packaging
   guidelines (@dwaring87, @jmh579, ropensci/software-review#792).
+* `brapi_get()` gains a `max_pages` argument. An unfiltered call against a
+  production server walks every page: Cassavabase holds 8,539 studies and
+  took 18 minutes to return them, and a germplasm listing on T3/Wheat ran
+  for over an hour before being interrupted. There was no way to ask for
+  just the first page. The default fetches everything as before, and a
+  truncated result is never cached.
 
 ### Deprecated
 
