### 2026-09-19 c07ab67
Stop naming unreleased packages as the downstream pipeline
DESIGN.md listed phenoQC, vcf2dosage, ggvariant and gsbench as the tools downstream of brapiR2. All four are mine; two are unreleased and two are at early versions, so the pipeline was described as though it were established. The scope boundary stands without them, grounded in the packages the articles actually hand output to.
