### 2026-09-20 7db1acd
Tidy NEWS formatting and record the DESIGN.md change
