### 2026-09-20 afe7b9e
Note that the token travels with the connection object
The vignette's credentials section covered how to keep a token out of scripts but not that it is then held in the connection object, so saveRDS() writes it to disk and str() prints it. Also bumps the R-CMD-check dependency cache, which held a corrupt knitr download on the macOS runner.
