+++ reference/brapi_cache_clear.md
@@ -25,9 +25,9 @@ Invisibly returns `con`.
 ``` r
 con <- brapi_connection("https://test-server.brapi.org")
 con <- brapi_cache_enable(con, dir = tempdir())
-#> ✔ Caching enabled at /tmp/RtmpEzSO8u (TTL: 3600s)
+#> ✔ Caching enabled at /tmp/RtmpiQhUre (TTL: 3600s)
 brapi_cache_clear(con)
-#> Warning: cannot remove file '/tmp/RtmpEzSO8u/bslib-e9b2b13fa612f50d23e4850d93d60d01', reason 'Directory not empty'
-#> Warning: cannot remove file '/tmp/RtmpEzSO8u/downlit', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/RtmpiQhUre/bslib-e9b2b13fa612f50d23e4850d93d60d01', reason 'Directory not empty'
+#> Warning: cannot remove file '/tmp/RtmpiQhUre/downlit', reason 'Directory not empty'
 #> ✔ Cleared 7 cached response(s).
 ```
