### 2026-09-20 62567de
Fetch packages from CRAN rather than RSPM
knitr 1.52 could not be extracted on the macOS runner across four consecutive runs, while every other dependency installed from the same source. RSPM's copy of that file is 85 kB larger than CRAN's and served as binary/octet-stream rather than application/x-gzip.
