### 2026-09-20 c1d6077
Route macOS binaries through RSPM to avoid zstd .tgz
CRAN's macOS builder now compresses binaries with zstd, which
pkgdepends does not recognise (r-lib/pkgdepends#485), so pak fails
with 'unknown archive type'. use-public-rspm: true applies only to
Linux; 'always' covers macOS, where RSPM still serves gzip.
