### 2026-09-23 a6d9270
Move to the rOpenSci organization
Repository URLs, badges and the user agent now point at ropensci/brapiR2.
Documentation moves to rOpenSci's central docs build, so the pkgdown
workflow and the template styling are gone and the site URL is
docs.ropensci.org/brapiR2. Adds the R-universe install method to the
README, switches the vignette's install instructions to match, and
releases 0.2.0 with the changes made during peer review.
