### 2026-09-24 8b420ff
Point the citation and package docs at the new repository
inst/CITATION had the owner and version written by hand, so it still
named the old repository and version 0.1.0. The version now comes from
DESCRIPTION. Regenerates brapiR2-package.Rd, which carried the old URLs.
