### 2026-09-24 61aac9d
Guard examples on the test server being reachable
Every example makes live requests to the public BrAPI test server, and
CRAN runs donttest examples on several check flavours - so a slow or
unavailable server would fail the check. Examples now run only if
brapi_ping() succeeds, which reports the server being unreachable
instead of erroring.
