+++ R/utils.R
@@ -69,7 +69,9 @@ brapi_ping <- function(con) {
 #' @examples
 #' \donttest{
 #' con <- brapi_connection("https://test-server.brapi.org")
-#' brapi_endpoints(con)
+#' if (brapi_ping(con)) {
+#'   brapi_endpoints(con)
+#' }
 #' }
 #'
 #' @export
