+++ R/auth.R
@@ -16,8 +16,10 @@
 #' @examples
 #' \donttest{
 #' con <- brapi_connection("https://test-server.brapi.org")
-#' con <- brapi_login(con, "brapi_reader", "brapi_reader")
-#' con
+#' if (brapi_ping(con)) {
+#'   con <- brapi_login(con, "brapi_reader", "brapi_reader")
+#'   con
+#' }
 #' }
 #'
 #' @export
@@ -76,14 +78,16 @@ brapi_login <- function(con, username, password) {
 #' @examples
 #' \donttest{
 #' con <- brapi_connection("https://test-server.brapi.org")
-#' con <- brapi_login_oauth2(
-#'   con,
-#'   client_id = "brapi_client",
-#'   client_secret = "brapi_secret",
-#'   authorize_url = "https://test-server.brapi.org/brapi/v2/authorize",
-#'   access_url = "https://test-server.brapi.org/brapi/v2/token"
-#' )
-#' con
+#' if (brapi_ping(con)) {
+#'   con <- brapi_login_oauth2(
+#'     con,
+#'     client_id = "brapi_client",
+#'     client_secret = "brapi_secret",
+#'     authorize_url = "https://test-server.brapi.org/brapi/v2/authorize",
+#'     access_url = "https://test-server.brapi.org/brapi/v2/token"
+#'   )
+#'   con
+#' }
 #' }
 #'
 #' @export

[…]
