### 2026-09-24 04f10ea
Prepare for CRAN submission
Adds cran-comments.md covering the title-case and spelling notes and how
examples and tests behave without a network. The citation URL lacked a
trailing slash and redirected, and the README's CONTRIBUTING.md link was
relative, so it pointed at nothing in the built tarball.
