+++ README.md
@@ -219,7 +219,8 @@ deliberately omits.
 
 ## Contributing
 
-Contributions are welcome! Please see [CONTRIBUTING.md](CONTRIBUTING.md)
+Contributions are welcome! Please see
+[CONTRIBUTING.md](https://github.com/ropensci/brapiR2/blob/main/CONTRIBUTING.md)
 for guidelines.
 
 ## License
